Move methods to blending.cpp/crafting.cpp
