+++ crafting.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+namespace elona
+{
+
+void initialize_craft_material_data();
+void crafting_menu();
+int get_required_craft_materials();
+int prodcheck();
+
+}
