+++ elonacore.cpp
@@ -2,6 +2,7 @@
 #include "animation.hpp"
 #include "audio.hpp"
 #include "autopick.hpp"
+#include "blending.hpp"
 #include "buff.hpp"
 #include "calc.hpp"
 #include "card.hpp"
@@ -9,6 +10,7 @@
 #include "class.hpp"
 #include "config.hpp"
 #include "ctrl_file.hpp"
+#include "crafting.hpp"
 #include "debug.hpp"
 #include "draw.hpp"
 #include "elona.hpp"
@@ -288,7 +290,6 @@ int prm_853;
 
 int tmp_at_m172;
 
-elona_vector1<int> inhlist_at_m184;
 elona_vector1<int> inhlist;
 
 int i_at_m68;
@@ -340,7 +341,6 @@ elona_vector2<int> egoenc;
 elona_vector2<int> egoref;
 elona_vector1<int> egolist;
 int maxegominorn = 0;
-elona_vector1<std::string> rpdatan;
 int i_at_m66 = 0;
 int lv_at_m77 = 0;
 int exp_at_m77 = 0;
@@ -375,12 +375,6 @@ int x2_at_m105 = 0;
 int y2_at_m105 = 0;
 int curmenu = 0;
 int inf_ver = 0;
-elona_vector1<int> matval;
-elona_vector2<int> matref;
-elona_vector2<int> matspot;
-elona_vector1<std::string> matname;
-elona_vector1<std::string> matdesc;
-int matid = 0;
 int characreatehack = 0;
 elona_vector1<std::string> usertxt;
 int p_at_m117 = 0;
@@ -475,7 +469,6 @@ elona_vector1<int> isetgod;
 elona_vector1<int> asettown;
 elona_vector1<int> moneybox;
 elona_vector1<int> isetbook;
-elona_vector1<int> rpsourcelist;
 int cibk = 0;
 int fdmax = 0;
 elona_vector2<int> fdlist;
@@ -542,7 +535,6 @@ int renewmulti = 0;
 int usermapid = 0;
 int n_at_m174 = 0;
 std::string s_at_m174;
-int prodtype = 0;
 int atxthrough = 0;
 std::string atxbg;
 std::string atxbgbk;
@@ -555,13 +547,6 @@ int atxap = 0;
 int atxlv = 0;
 int txtadvscreenupdate = 0;
 int dbg_skipevent = 0;
-elona_vector2<int> rpdata;
-int rpid = 0;
-elona_vector2<std::string> rfnameorg;
-elona_vector1<int> rpref;
-int rpmode = 0;
-elona_vector1<int> rppage;
-int rpresult = 0;
 int cansee = 0;
 int kdx = 0;
 int kdy = 0;
@@ -5029,22 +5014,6 @@ void set_material_specific_attributes()
 
 
 
-std::string rpname(int prm_516)
-{
-    std::string s_at_m62;
-    s_at_m62 = u8"?????"s;
-    if (rpdatan(prm_516) != ""s)
-    {
-        return rpdatan(prm_516);
-    }
-    if (prm_516 >= 400)
-    {
-        return ioriginalnameref(prm_516 - 400);
-    }
-    return s_at_m62;
-}
-
-
 
 void colorres(int)
 {
@@ -8038,1040 +8007,6 @@ int direction(int x1, int y1, int x2, int y2)
     }
 }
 
-
-
-void initialize_craft_material_data()
-{
-    DIM2(matval, 4);
-    DIM3(matref, 5, 400);
-    DIM3(matspot, 5, 400);
-    SDIM3(matname, 40, 400);
-    SDIM3(matdesc, 80, 400);
-
-    matref(0, 0) = 0;
-    matref(1, 0) = 1;
-    matref(2, 0) = 240;
-    matname(0) = lang(u8"クズ"s, u8"Garbage"s);
-    matdesc(0) = lang(u8"何の価値もないクズ"s, u8"Just garbage."s);
-    matspot(0, 0) = 18;
-
-    matref(0, 1) = 1;
-    matref(1, 1) = 80;
-    matref(2, 1) = 434;
-    matname(1) = lang(u8"カジノチップ"s, u8"Casino chip"s);
-    matdesc(1) = lang(u8"カジノで使うチップ"s, u8"Chips used in casinos."s);
-    matspot(0, 1) = 18;
-
-    matref(0, 40) = 1;
-    matref(1, 40) = 200;
-    matref(2, 40) = 437;
-    matname(40) = lang(u8"100Yen硬貨"s, u8"100 Yen coin"s);
-    matdesc(40) = lang(u8"異国のお金"s, u8"Foreign money."s);
-    matspot(0, 40) = 18;
-
-    matref(0, 41) = 1;
-    matref(1, 41) = 1000;
-    matref(2, 41) = 437;
-    matname(41) = lang(u8"500Yen硬貨"s, u8"500 Yen coin"s);
-    matdesc(41) = lang(u8"異国のお金"s, u8"Foreign money."s);
-    matspot(0, 41) = 18;
-
-    matref(0, 43) = 1;
-    matref(1, 43) = 20;
-    matref(2, 43) = 470;
-    matname(43) = lang(u8"紙"s, u8"Paper"s);
-    matdesc(43) = lang(u8"白紙"s, u8"Blank paper."s);
-    matspot(0, 43) = 18;
-
-    matref(0, 2) = 1;
-    matref(1, 2) = 10;
-    matref(2, 2) = 239;
-    matname(2) = lang(u8"炭"s, u8"Charcoal"s);
-    matdesc(2) = lang(u8"良質な炭"s, u8"This is good quality charcoal."s);
-    matspot(0, 2) = 18;
-
-    matref(0, 3) = 1;
-    matref(1, 3) = 10;
-    matref(2, 3) = 239;
-    matname(3) = lang(u8"流木"s, u8"Driftwood"s);
-    matdesc(3) = lang(u8"あまり質の良くない木材"s, u8"Common driftwood."s);
-    matspot(0, 3) = 18;
-
-    matref(0, 48) = 1;
-    matref(1, 48) = 10;
-    matref(2, 48) = 209;
-    matname(48) = lang(u8"石ころ"s, u8"Pebble"s);
-    matdesc(48) =
-        lang(u8"どこにでも落ちている"s, u8"These are around everywhere."s);
-    matspot(0, 48) = 18;
-
-    matref(0, 6) = 1;
-    matref(1, 6) = 70;
-    matref(2, 6) = 239;
-    matname(6) = lang(u8"杖"s, u8"Stick"s);
-    matdesc(6) = lang(u8"杖"s, u8"A specially shaped piece of wood."s);
-    matspot(0, 6) = 18;
-
-    matref(0, 45) = 3;
-    matref(1, 45) = 20;
-    matref(2, 45) = 155;
-    matname(45) = lang(u8"布きれ"s, u8"Cloth"s);
-    matdesc(45) = lang(u8"ただの布きれ"s, u8"A piece of cloth."s);
-    matspot(0, 45) = 18;
-
-    matref(0, 38) = 4;
-    matref(1, 38) = 30;
-    matref(2, 38) = 8;
-    matname(38) = lang(u8"わめく狂人"s, u8"Yelling madman"s);
-    matdesc(38) = lang(u8"黙らない"s, u8"It never shuts up."s);
-    matspot(0, 38) = 18;
-
-    matref(0, 30) = 10;
-    matref(1, 30) = 150;
-    matref(2, 30) = 7;
-    matname(30) = lang(u8"マジックマス"s, u8"Magic mass"s);
-    matdesc(30) = lang(u8"謎の魔力の塊"s, u8"Unknown magical mass."s);
-    matspot(0, 30) = 18;
-
-    matref(0, 23) = 12;
-    matref(1, 23) = 20;
-    matref(2, 23) = 7;
-    matname(23) = lang(u8"電気"s, u8"Electricity"s);
-    matdesc(23) = lang(u8"放電している"s, u8"It is discharging."s);
-    matspot(0, 23) = 18;
-
-    matref(0, 44) = 25;
-    matref(1, 44) = 150;
-    matref(2, 44) = 7;
-    matname(44) = lang(u8"生成機械"s, u8"Generator"s);
-    matdesc(44) =
-        lang(u8"何かを生み出す機械"s, u8"This machine generates something."s);
-    matspot(0, 44) = 18;
-
-    matref(0, 50) = 1;
-    matref(1, 50) = 30;
-    matref(2, 50) = 209;
-    matname(50) = lang(u8"魔力の結晶"s, u8"Magic fragment"s);
-    matdesc(50) = lang(u8"魔力が込められている"s, u8"This is magical."s);
-    matspot(0, 50) = 9;
-
-    matref(0, 46) = 2;
-    matref(1, 46) = 20;
-    matref(2, 46) = 239;
-    matname(46) = lang(u8"枝"s, u8"Branch"s);
-    matdesc(46) = lang(u8"枝"s, u8"Broken off of a tree."s);
-    matspot(0, 46) = 10;
-
-    matref(0, 33) = 4;
-    matref(1, 33) = 20;
-    matref(2, 33) = 155;
-    matname(33) = lang(u8"革"s, u8"Leather"s);
-    matdesc(33) = lang(u8"なめした皮"s, u8"A tanned hide."s);
-    matspot(0, 33) = 10;
-
-    matref(0, 53) = 6;
-    matref(1, 53) = 20;
-    matref(2, 53) = 202;
-    matname(53) = lang(u8"ツル"s, u8"Vein"s);
-    matdesc(53) = lang(u8"ツル"s, u8"Vein."s);
-    matspot(0, 53) = 10;
-
-    matref(0, 47) = 10;
-    matref(1, 47) = 20;
-    matref(2, 47) = 239;
-    matname(47) = lang(u8"頑丈な枝"s, u8"Thick wood"s);
-    matdesc(47) =
-        lang(u8"なかなか折れない"s, u8"This doesn't break down easily."s);
-    matspot(0, 47) = 10;
-
-    matref(0, 37) = 12;
-    matref(1, 37) = 60;
-    matref(2, 37) = 471;
-    matname(37) = lang(u8"曲がった杖"s, u8"Curved stick"s);
-    matdesc(37) = lang(u8"杖の役割を果たさない"s, u8"Almost unusable."s);
-    matspot(0, 37) = 10;
-
-    matref(0, 54) = 1;
-    matref(1, 54) = 50;
-    matref(2, 54) = 354;
-    matname(54) = lang(u8"接着剤"s, u8"Adhesive"s);
-    matdesc(54) = lang(u8"ネトネトしている"s, u8"It's sticky."s);
-    matspot(0, 54) = 12;
-
-    matref(0, 49) = 5;
-    matref(1, 49) = 50;
-    matref(2, 49) = 150;
-    matname(49) = lang(u8"記憶の欠片"s, u8"Memory fragment"s);
-    matdesc(49) =
-        lang(u8"思い出が残っている"s, u8" Reminds you of something."s);
-    matspot(0, 49) = 12;
-
-    matref(0, 35) = 8;
-    matref(1, 35) = 40;
-    matref(2, 35) = 470;
-    matname(35) = lang(u8"魔法の紙"s, u8"Magic paper"s);
-    matdesc(35) = lang(u8"魔力を秘めた紙"s, u8"Magical paper."s);
-    matspot(0, 35) = 12;
-
-    matref(0, 36) = 8;
-    matref(1, 36) = 40;
-    matref(2, 36) = 354;
-    matname(36) = lang(u8"魔法のインク"s, u8"Magic ink"s);
-    matdesc(36) = lang(u8"魔力を秘めたインク"s, u8"Wizards love to use it."s);
-    matspot(0, 36) = 12;
-
-    matref(0, 12) = 1;
-    matref(1, 12) = 10;
-    matref(2, 12) = 354;
-    matname(12) = lang(u8"海水"s, u8"Seawater"s);
-    matdesc(12) = lang(u8"しょっぱい水"s, u8"Very salty."s);
-    matspot(0, 12) = 13;
-
-    matref(0, 5) = 1;
-    matref(1, 5) = 10;
-    matref(2, 5) = 354;
-    matname(5) = lang(u8"水滴"s, u8"Waterdrop"s);
-    matdesc(5) = lang(u8"水滴"s, u8"A drop of water."s);
-    matspot(0, 5) = 13;
-
-    matref(0, 10) = 4;
-    matref(1, 10) = 50;
-    matref(2, 10) = 354;
-    matname(10) = lang(u8"天使の涙"s, u8"Angel's tear"s);
-    matdesc(10) = lang(u8"不思議な輝きを放つ涙"s, u8"Hard to gather."s);
-    matspot(0, 10) = 13;
-
-    matref(0, 25) = 4;
-    matref(1, 25) = 20;
-    matref(2, 25) = 354;
-    matname(25) = lang(u8"お湯"s, u8"Hot water"s);
-    matdesc(25) =
-        lang(u8"なぜか冷めない"s, u8"Somehow it's not cooling down."s);
-    matspot(0, 25) = 13;
-
-    matref(0, 11) = 8;
-    matref(1, 11) = 30;
-    matref(2, 11) = 354;
-    matname(11) = lang(u8"魔女の涙"s, u8"Witch's tear"s);
-    matdesc(11) = lang(u8"決して乾かない涙"s, u8"Never dries up."s);
-    matspot(0, 11) = 13;
-
-    matref(0, 20) = 12;
-    matref(1, 20) = 15;
-    matref(2, 20) = 7;
-    matname(20) = lang(u8"雪"s, u8"Snow"s);
-    matdesc(20) = lang(u8"なぜか溶けない"s, u8"It won't melt somehow."s);
-    matspot(0, 20) = 13;
-
-    matref(0, 7) = 1;
-    matref(1, 7) = 40;
-    matref(2, 7) = 209;
-    matname(7) = lang(u8"ミスリルの欠片"s, u8"Mithril fragment"s);
-    matdesc(7) = lang(u8"ミスリルを含んだ鉱石の欠片"s, u8"Contains mithril."s);
-    matspot(0, 7) = 14;
-
-    matref(0, 9) = 1;
-    matref(1, 9) = 10;
-    matref(2, 9) = 209;
-    matname(9) = lang(u8"鉄の欠片"s, u8"Iron fragment"s);
-    matdesc(9) = lang(u8"鉄分を含んだ鉱石の欠片"s, u8"Contains iron."s);
-    matspot(0, 9) = 14;
-
-    matref(0, 26) = 5;
-    matref(1, 26) = 10;
-    matref(2, 26) = 209;
-    matname(26) = lang(u8"火炎石"s, u8"Fire stone"s);
-    matdesc(26) = lang(u8"熱を帯びた石"s, u8"This is hot."s);
-    matspot(0, 26) = 14;
-
-    matref(0, 27) = 5;
-    matref(1, 27) = 10;
-    matref(2, 27) = 209;
-    matname(27) = lang(u8"冷凍石"s, u8"Ice stone"s);
-    matdesc(27) = lang(u8"冷たい石"s, u8"This is cool."s);
-    matspot(0, 27) = 14;
-
-    matref(0, 28) = 5;
-    matref(1, 28) = 10;
-    matref(2, 28) = 209;
-    matname(28) = lang(u8"帯電石"s, u8"Discharging stone"s);
-    matdesc(28) = lang(u8"電気を帯びた石"s, u8"It's electrified."s);
-    matspot(0, 28) = 14;
-
-    matref(0, 52) = 8;
-    matref(1, 52) = 50;
-    matref(2, 52) = 209;
-    matname(52) = lang(u8"上等な石"s, u8"Fine stone"s);
-    matdesc(52) = lang(u8"状態の良い石"s, u8"A good quality stone."s);
-    matspot(0, 52) = 14;
-
-    matref(0, 8) = 10;
-    matref(1, 8) = 40;
-    matref(2, 8) = 209;
-    matname(8) = lang(u8"エーテルの欠片"s, u8"Ether fragment"s);
-    matdesc(8) = lang(u8"エーテルを含んだ鉱石の欠片"s, u8"Contains ether."s);
-    matspot(0, 8) = 14;
-
-    matref(0, 22) = 20;
-    matref(1, 22) = 100;
-    matref(2, 22) = 209;
-    matname(22) = lang(u8"元素の欠片"s, u8"Element fragment"s);
-    matdesc(22) = lang(u8"元素の力を宿した石"s, u8"Elemental."s);
-    matspot(0, 22) = 14;
-
-    matref(0, 51) = 40;
-    matref(1, 51) = 200;
-    matref(2, 51) = 209;
-    matname(51) = lang(u8"混沌の石"s, u8"Chaos stone"s);
-    matdesc(51) = lang(u8"混沌の力が封じられた石"s, u8"Chaotic."s);
-    matspot(0, 51) = 14;
-
-    matref(0, 13) = 1;
-    matref(1, 13) = 10;
-    matref(2, 13) = 170;
-    matname(13) = lang(u8"唸る草"s, u8"Howling weed"s);
-    matdesc(13) = lang(u8"変な音を出す草"s, u8"It's making a whining sound."s);
-    matspot(0, 13) = 15;
-
-    matref(0, 14) = 1;
-    matref(1, 14) = 20;
-    matref(2, 14) = 170;
-    matname(14) = lang(u8"赤い草"s, u8"Red weed"s);
-    matdesc(14) = lang(u8"ほんのり赤い草"s, u8"A faintly reddish weed."s);
-    matspot(0, 14) = 15;
-
-    matref(0, 15) = 1;
-    matref(1, 15) = 20;
-    matref(2, 15) = 170;
-    matname(15) = lang(u8"青い草"s, u8"Blue weed"s);
-    matdesc(15) = lang(u8"ほんのり青い草"s, u8"A faintly blueish weed."s);
-    matspot(0, 15) = 15;
-
-    matref(0, 16) = 1;
-    matref(1, 16) = 10;
-    matref(2, 16) = 170;
-    matname(16) = lang(u8"呪いの草"s, u8"Curse weed"s);
-    matdesc(16) = lang(u8"呪われた草"s, u8"A cursed weed."s);
-    matspot(0, 16) = 15;
-
-    matref(0, 42) = 3;
-    matref(1, 42) = 20;
-    matref(2, 42) = 170;
-    matname(42) = lang(u8"薬草"s, u8"Medicinal weed"s);
-    matdesc(42) = lang(u8"少し傷を癒す"s, u8"Slightly cures wounds."s);
-    matspot(0, 42) = 15;
-
-    matref(0, 29) = 7;
-    matref(1, 29) = 50;
-    matref(2, 29) = 170;
-    matname(29) = lang(u8"浮き草"s, u8"Floating weed"s);
-    matdesc(29) = lang(u8"とても軽い草"s, u8"Very light."s);
-    matspot(0, 29) = 15;
-
-    matref(0, 17) = 12;
-    matref(1, 17) = 20;
-    matref(2, 17) = 170;
-    matname(17) = lang(u8"聖なる草"s, u8"Holy weed"s);
-    matdesc(17) = lang(u8"祝福された草"s, u8"It's blessed."s);
-    matspot(0, 17) = 15;
-
-    matref(0, 24) = 15;
-    matref(1, 24) = 40;
-    matref(2, 24) = 8;
-    matname(24) = lang(u8"黒い霧"s, u8"Black mist"s);
-    matdesc(24) = lang(u8"真っ黒な霧"s, u8"It's as black as coal."s);
-    matspot(0, 24) = 15;
-
-    matref(0, 34) = 25;
-    matref(1, 34) = 200;
-    matref(2, 34) = 354;
-    matname(34) = lang(u8"世界樹の樹液"s, u8"Sap of Yaggdrasil"s);
-    matdesc(34) = lang(u8"どんな傷も癒す樹液"s, u8"It's pretty rare."s);
-    matspot(0, 34) = 15;
-
-    matspot(1, 34) = 10;
-    matref(0, 4) = 1;
-    matref(1, 4) = 10;
-    matref(2, 4) = 436;
-    matname(4) = lang(u8"鳥の羽"s, u8"Bird's feather"s);
-    matdesc(4) = lang(u8"普通の鳥の羽"s, u8"A piece of Feather."s);
-    matspot(0, 4) = 16;
-
-    matref(0, 18) = 1;
-    matref(1, 18) = 80;
-    matref(2, 18) = 301;
-    matname(18) = lang(u8"ウサギの尻尾"s, u8"Rabbit's tail"s);
-    matdesc(18) = lang(u8"幸運を呼ぶ尻尾"s, u8"Scut."s);
-    matspot(0, 18) = 16;
-
-    matref(0, 31) = 1;
-    matref(1, 31) = 20;
-    matref(2, 31) = 304;
-    matname(31) = lang(u8"ヒトの遺伝子"s, u8"Human gene"s);
-    matdesc(31) = lang(u8"普通の遺伝子"s, u8"It's common."s);
-    matspot(0, 31) = 16;
-
-    matref(0, 39) = 3;
-    matref(1, 39) = 20;
-    matref(2, 39) = 301;
-    matname(39) = lang(u8"クマの尻尾"s, u8"Bear's tail"s);
-    matdesc(39) = lang(u8"とても短い"s, u8"Very short."s);
-    matspot(0, 39) = 16;
-
-    matref(0, 19) = 5;
-    matref(1, 19) = 15;
-    matref(2, 19) = 304;
-    matname(19) = lang(u8"トロールの遺伝子"s, u8"Troll gene"s);
-    matdesc(19) = lang(u8"再生する遺伝子"s, u8"It's regenerating."s);
-    matspot(0, 19) = 16;
-
-    matref(0, 32) = 8;
-    matref(1, 32) = 40;
-    matref(2, 32) = 304;
-    matname(32) = lang(u8"魔女の瞳"s, u8"Witch's eye"s);
-    matdesc(32) = lang(u8"まだ動いている"s, u8"These eyes are still moving."s);
-    matspot(0, 32) = 16;
-
-    matref(0, 21) = 15;
-    matref(1, 21) = 45;
-    matref(2, 21) = 209;
-    matname(21) = lang(u8"フェアリーダスト"s, u8"Fairy dust"s);
-    matdesc(21) = lang(u8"妖精の燐粉"s, u8"Fairies leave it."s);
-    matspot(0, 21) = 16;
-}
-
-
-
-int prodcheck()
-{
-    int f_at_m110 = 0;
-    elona_vector1<int> j_at_m110;
-    f_at_m110 = 1;
-    if (matval(1) > sdata(matval, 0))
-    {
-        f_at_m110 = -1;
-    }
-    for (int cnt = 0; cnt < 6; ++cnt)
-    {
-        j_at_m110(0) = matneed(cnt * 2);
-        j_at_m110(1) = matneed(cnt * 2 + 1);
-        if (j_at_m110 == -1)
-        {
-            break;
-        }
-        if (mat(j_at_m110) < j_at_m110(1))
-        {
-            f_at_m110 = -1;
-        }
-    }
-    return f_at_m110;
-}
-
-
-
-int get_required_craft_materials()
-{
-    matneed(0) = -1;
-    matneed(1) = -1;
-    matneed(2) = -1;
-    matneed(3) = -1;
-    matneed(4) = -1;
-    matneed(5) = -1;
-    matneed(6) = -1;
-    matneed(7) = -1;
-    matneed(8) = -1;
-    matneed(9) = -1;
-    if (matid == 502)
-    {
-        matval(0) = 179;
-        matval(1) = 40;
-        matval(2) = 100;
-        matneed(0) = 35;
-        matneed(1) = 2;
-        matneed(2) = 44;
-        matneed(3) = 2;
-        matneed(4) = 51;
-        matneed(5) = 2;
-        return 1;
-    }
-    if (matid == 501)
-    {
-        matval(0) = 179;
-        matval(1) = 13;
-        matval(2) = 100;
-        matneed(0) = 43;
-        matneed(1) = 2;
-        matneed(2) = 44;
-        matneed(3) = 1;
-        return 1;
-    }
-    if (matid == 500)
-    {
-        matval(0) = 179;
-        matval(1) = 1;
-        matval(2) = 100;
-        matneed(0) = 43;
-        matneed(1) = 2;
-        matneed(2) = 38;
-        matneed(3) = 1;
-        return 1;
-    }
-    if (matid == 455)
-    {
-        matval(0) = 177;
-        matval(1) = 1;
-        matval(2) = 100;
-        matneed(0) = 45;
-        matneed(1) = 4;
-        matneed(2) = 53;
-        matneed(3) = 2;
-        matneed(4) = 7;
-        matneed(5) = 3;
-        return 1;
-    }
-    if (matid == 411)
-    {
-        matval(0) = 179;
-        matval(1) = 5;
-        matval(2) = 100;
-        matneed(0) = 43;
-        matneed(1) = 4;
-        matneed(2) = 36;
-        matneed(3) = 1;
-        matneed(4) = 23;
-        matneed(5) = 2;
-        return 1;
-    }
-    if (matid == 408)
-    {
-        matval(0) = 176;
-        matval(1) = 11;
-        matval(2) = 100;
-        matneed(0) = 47;
-        matneed(1) = 3;
-        matneed(2) = 52;
-        matneed(3) = 2;
-        matneed(4) = 21;
-        matneed(5) = 1;
-        return 1;
-    }
-    if (matid == 407)
-    {
-        matval(0) = 178;
-        matval(1) = 14;
-        matval(2) = 100;
-        matneed(0) = 25;
-        matneed(1) = 4;
-        matneed(2) = 20;
-        matneed(3) = 3;
-        matneed(4) = 42;
-        matneed(5) = 1;
-        return 1;
-    }
-    if (matid == 405)
-    {
-        matval(0) = 176;
-        matval(1) = 15;
-        matval(2) = 100;
-        matneed(0) = 48;
-        matneed(1) = 8;
-        matneed(2) = 9;
-        matneed(3) = 5;
-        matneed(4) = 52;
-        matneed(5) = 2;
-        return 1;
-    }
-    if (matid == 403)
-    {
-        matval(0) = 176;
-        matval(1) = 12;
-        matval(2) = 100;
-        matneed(0) = 47;
-        matneed(1) = 5;
-        matneed(2) = 16;
-        matneed(3) = 2;
-        matneed(4) = 10;
-        matneed(5) = 1;
-        return 1;
-    }
-    if (matid == 402)
-    {
-        matval(0) = 177;
-        matval(1) = 5;
-        matval(2) = 100;
-        matneed(0) = 53;
-        matneed(1) = 7;
-        matneed(2) = 45;
-        matneed(3) = 4;
-        return 1;
-    }
-    if (matid == 401)
-    {
-        matval(0) = 179;
-        matval(1) = 14;
-        matval(2) = 100;
-        matneed(0) = 3;
-        matneed(1) = 8;
-        matneed(2) = 53;
-        matneed(3) = 4;
-        matneed(4) = 21;
-        matneed(5) = 1;
-        return 1;
-    }
-    if (matid == 400)
-    {
-        matval(0) = 176;
-        matval(1) = 8;
-        matval(2) = 100;
-        matneed(0) = 3;
-        matneed(1) = 6;
-        matneed(2) = 46;
-        matneed(3) = 4;
-        matneed(4) = 2;
-        matneed(5) = 4;
-        return 1;
-    }
-    if (matid == 399)
-    {
-        matval(0) = 177;
-        matval(1) = 8;
-        matval(2) = 100;
-        matneed(0) = 45;
-        matneed(1) = 5;
-        matneed(2) = 33;
-        matneed(3) = 2;
-        matneed(4) = 20;
-        matneed(5) = 2;
-        return 1;
-    }
-    if (matid == 395)
-    {
-        matval(0) = 179;
-        matval(1) = 30;
-        matval(2) = 100;
-        matneed(0) = 43;
-        matneed(1) = 5;
-        matneed(2) = 30;
-        matneed(3) = 1;
-        matneed(4) = 44;
-        matneed(5) = 1;
-        return 1;
-    }
-    if (matid == 390)
-    {
-        matval(0) = 179;
-        matval(1) = 25;
-        matval(2) = 100;
-        matneed(0) = 35;
-        matneed(1) = 1;
-        matneed(2) = 36;
-        matneed(3) = 2;
-        matneed(4) = 30;
-        matneed(5) = 2;
-        return 1;
-    }
-    if (matid == 389)
-    {
-        matval(0) = 179;
-        matval(1) = 25;
-        matval(2) = 100;
-        matneed(0) = 35;
-        matneed(1) = 2;
-        matneed(2) = 36;
-        matneed(3) = 2;
-        matneed(4) = 17;
-        matneed(5) = 4;
-        return 1;
-    }
-    if (matid == 388)
-    {
-        matval(0) = 179;
-        matval(1) = 15;
-        matval(2) = 100;
-        matneed(0) = 35;
-        matneed(1) = 1;
-        matneed(2) = 17;
-        matneed(3) = 2;
-        matneed(4) = 23;
-        matneed(5) = 3;
-        return 1;
-    }
-    if (matid == 384)
-    {
-        matval(0) = 179;
-        matval(1) = 15;
-        matval(2) = 100;
-        matneed(0) = 35;
-        matneed(1) = 2;
-        matneed(2) = 17;
-        matneed(3) = 4;
-        matneed(4) = 16;
-        matneed(5) = 4;
-        return 1;
-    }
-    if (matid == 362)
-    {
-        matval(0) = 179;
-        matval(1) = 18;
-        matval(2) = 100;
-        matneed(0) = 35;
-        matneed(1) = 1;
-        matneed(2) = 49;
-        matneed(3) = 2;
-        matneed(4) = 30;
-        matneed(5) = 1;
-        return 1;
-    }
-    if (matid == 287)
-    {
-        matval(0) = 178;
-        matval(1) = 45;
-        matval(2) = 100;
-        matneed(0) = 34;
-        matneed(1) = 2;
-        matneed(2) = 10;
-        matneed(3) = 5;
-        matneed(4) = 25;
-        matneed(5) = 4;
-        return 1;
-    }
-    if (matid == 286)
-    {
-        matval(0) = 178;
-        matval(1) = 8;
-        matval(2) = 100;
-        matneed(0) = 25;
-        matneed(1) = 4;
-        matneed(2) = 14;
-        matneed(3) = 4;
-        return 1;
-    }
-    if (matid == 285)
-    {
-        matval(0) = 178;
-        matval(1) = 8;
-        matval(2) = 100;
-        matneed(0) = 25;
-        matneed(1) = 4;
-        matneed(2) = 15;
-        matneed(3) = 4;
-        return 1;
-    }
-    if (matid == 253)
-    {
-        matval(0) = 178;
-        matval(1) = 1;
-        matval(2) = 100;
-        matneed(0) = 5;
-        matneed(1) = 4;
-        return 1;
-    }
-    if (matid == 245)
-    {
-        matval(0) = 179;
-        matval(1) = 1;
-        matval(2) = 100;
-        matneed(0) = 43;
-        matneed(1) = 2;
-        matneed(2) = 2;
-        matneed(3) = 2;
-        matneed(4) = 4;
-        matneed(5) = 2;
-        return 1;
-    }
-    if (matid == 242)
-    {
-        matval(0) = 179;
-        matval(1) = 10;
-        matval(2) = 100;
-        matneed(0) = 35;
-        matneed(1) = 1;
-        matneed(2) = 36;
-        matneed(3) = 1;
-        matneed(4) = 23;
-        matneed(5) = 3;
-        return 1;
-    }
-    if (matid == 236)
-    {
-        matval(0) = 179;
-        matval(1) = 4;
-        matval(2) = 100;
-        matneed(0) = 43;
-        matneed(1) = 3;
-        matneed(2) = 13;
-        matneed(3) = 2;
-        matneed(4) = 2;
-        matneed(5) = 2;
-        return 1;
-    }
-    if (matid == 209)
-    {
-        matval(0) = 179;
-        matval(1) = 8;
-        matval(2) = 100;
-        matneed(0) = 43;
-        matneed(1) = 4;
-        matneed(2) = 17;
-        matneed(3) = 2;
-        matneed(4) = 16;
-        matneed(5) = 2;
-        return 1;
-    }
-    if (matid == 175)
-    {
-        matval(0) = 176;
-        matval(1) = 17;
-        matval(2) = 100;
-        matneed(0) = 37;
-        matneed(1) = 1;
-        matneed(2) = 28;
-        matneed(3) = 5;
-        matneed(4) = 54;
-        matneed(5) = 1;
-        return 1;
-    }
-    if (matid == 123)
-    {
-        matval(0) = 176;
-        matval(1) = 14;
-        matval(2) = 100;
-        matneed(0) = 37;
-        matneed(1) = 1;
-        matneed(2) = 26;
-        matneed(3) = 5;
-        matneed(4) = 50;
-        matneed(5) = 2;
-        return 1;
-    }
-    if (matid == 122)
-    {
-        matval(0) = 176;
-        matval(1) = 14;
-        matval(2) = 100;
-        matneed(0) = 37;
-        matneed(1) = 1;
-        matneed(2) = 27;
-        matneed(3) = 5;
-        matneed(4) = 20;
-        matneed(5) = 5;
-        return 1;
-    }
-    if (matid == 120)
-    {
-        matval(0) = 176;
-        matval(1) = 1;
-        matval(2) = 100;
-        matneed(0) = 6;
-        matneed(1) = 1;
-        matneed(2) = 9;
-        matneed(3) = 4;
-        matneed(4) = 50;
-        matneed(5) = 2;
-        return 1;
-    }
-    if (matid == 119)
-    {
-        matval(0) = 176;
-        matval(1) = 1;
-        matval(2) = 100;
-        matneed(0) = 6;
-        matneed(1) = 1;
-        matneed(2) = 14;
-        matneed(3) = 4;
-        matneed(4) = 54;
-        matneed(5) = 1;
-        return 1;
-    }
-    if (matid == 76)
-    {
-        matval(0) = 178;
-        matval(1) = 40;
-        matval(2) = 100;
-        matneed(0) = 34;
-        matneed(1) = 2;
-        matneed(2) = 5;
-        matneed(3) = 5;
-        matneed(4) = 42;
-        matneed(5) = 4;
-        return 1;
-    }
-    if (matid == 75)
-    {
-        matval(0) = 178;
-        matval(1) = 30;
-        matval(2) = 100;
-        matneed(0) = 34;
-        matneed(1) = 1;
-        matneed(2) = 25;
-        matneed(3) = 4;
-        matneed(4) = 14;
-        matneed(5) = 5;
-        return 1;
-    }
-    if (matid == 74)
-    {
-        matval(0) = 178;
-        matval(1) = 25;
-        matval(2) = 100;
-        matneed(0) = 15;
-        matneed(1) = 3;
-        matneed(2) = 25;
-        matneed(3) = 4;
-        matneed(4) = 42;
-        matneed(5) = 3;
-        return 1;
-    }
-    if (matid == 72)
-    {
-        matval(0) = 178;
-        matval(1) = 20;
-        matval(2) = 100;
-        matneed(0) = 10;
-        matneed(1) = 1;
-        matneed(2) = 5;
-        matneed(3) = 4;
-        matneed(4) = 42;
-        matneed(5) = 2;
-        return 1;
-    }
-    if (matid == 71)
-    {
-        matval(0) = 178;
-        matval(1) = 15;
-        matval(2) = 100;
-        matneed(0) = 25;
-        matneed(1) = 3;
-        matneed(2) = 42;
-        matneed(3) = 2;
-        return 1;
-    }
-    if (matid == 70)
-    {
-        matval(0) = 178;
-        matval(1) = 10;
-        matval(2) = 100;
-        matneed(0) = 25;
-        matneed(1) = 3;
-        matneed(2) = 14;
-        matneed(3) = 1;
-        matneed(4) = 15;
-        matneed(5) = 1;
-        return 1;
-    }
-    if (matid == 69)
-    {
-        matval(0) = 178;
-        matval(1) = 5;
-        matval(2) = 100;
-        matneed(0) = 5;
-        matneed(1) = 3;
-        matneed(2) = 15;
-        matneed(3) = 2;
-        return 1;
-    }
-    if (matid == 68)
-    {
-        matval(0) = 178;
-        matval(1) = 1;
-        matval(2) = 100;
-        matneed(0) = 5;
-        matneed(1) = 3;
-        matneed(2) = 14;
-        matneed(3) = 1;
-        return 1;
-    }
-    if (matid == 66)
-    {
-        matval(0) = 177;
-        matval(1) = 8;
-        matval(2) = 100;
-        matneed(0) = 33;
-        matneed(1) = 4;
-        matneed(2) = 39;
-        matneed(3) = 5;
-        matneed(4) = 18;
-        matneed(5) = 2;
-        return 1;
-    }
-    if (matid == 65)
-    {
-        matval(0) = 177;
-        matval(1) = 10;
-        matval(2) = 100;
-        matneed(0) = 33;
-        matneed(1) = 5;
-        matneed(2) = 54;
-        matneed(3) = 3;
-        matneed(4) = 31;
-        matneed(5) = 6;
-        return 1;
-    }
-    if (matid == 19)
-    {
-        matval(0) = 176;
-        matval(1) = 5;
-        matval(2) = 100;
-        matneed(0) = 6;
-        matneed(1) = 1;
-        matneed(2) = 4;
-        matneed(3) = 4;
-        matneed(4) = 29;
-        matneed(5) = 2;
-        return 1;
-    }
-    if (matid == 18)
-    {
-        matval(0) = 176;
-        matval(1) = 8;
-        matval(2) = 100;
-        matneed(0) = 6;
-        matneed(1) = 1;
-        matneed(2) = 32;
-        matneed(3) = 2;
-        matneed(4) = 49;
-        matneed(5) = 2;
-        return 1;
-    }
-    if (matid == 16)
-    {
-        matval(0) = 179;
-        matval(1) = 7;
-        matval(2) = 100;
-        matneed(0) = 43;
-        matneed(1) = 3;
-        matneed(2) = 2;
-        matneed(3) = 2;
-        matneed(4) = 4;
-        matneed(5) = 3;
-        return 1;
-    }
-    if (matid == 14)
-    {
-        matval(0) = 179;
-        matval(1) = 1;
-        matval(2) = 100;
-        matneed(0) = 43;
-        matneed(1) = 2;
-        matneed(2) = 49;
-        matneed(3) = 1;
-        return 1;
-    }
-    if (matid == 8)
-    {
-        matval(0) = 177;
-        matval(1) = 4;
-        matval(2) = 100;
-        matneed(0) = 45;
-        matneed(1) = 6;
-        matneed(2) = 53;
-        matneed(3) = 3;
-        matneed(4) = 50;
-        matneed(5) = 2;
-        return 1;
-    }
-    return -1;
-}
 
 
 
@@ -32676,296 +31611,6 @@ void matdelmain(int prm_1033, int prm_1034)
 
 
 
-void label_18552()
-{
-    int matuse = 0;
-label_18551_internal:
-    listmax = 0;
-    page = 0;
-    pagesize = 10;
-    cs = 0;
-    cc = 0;
-    cs_bk = -1;
-    page_load();
-    if (invctrl == 0)
-    {
-        for (int cnt = 0, cnt_end = (maxitemid); cnt < cnt_end; ++cnt)
-        {
-            matid = cnt;
-            int stat = get_required_craft_materials();
-            if (stat == -1)
-            {
-                continue;
-            }
-            if (prodtype == 2)
-            {
-                if (matval != 178)
-                {
-                    continue;
-                }
-            }
-            if (prodtype == 1)
-            {
-                if (matval != 176)
-                {
-                    continue;
-                }
-            }
-            if (prodtype == 3)
-            {
-                if (matval != 179)
-                {
-                    continue;
-                }
-            }
-            if (prodtype == 4)
-            {
-                if (matval != 177)
-                {
-                    continue;
-                }
-            }
-            if (sdata(matval, 0) + 3 < matval(1))
-            {
-                continue;
-            }
-            listn(0, listmax) = ""s + prodcheck();
-            list(0, listmax) = cnt;
-            list(1, listmax) = 0;
-            ++listmax;
-        }
-    }
-    else
-    {
-        for (int cnt = 0; cnt < 50; ++cnt)
-        {
-            listn(0, listmax) = ""s + prodcheck();
-            list(0, listmax) = 630;
-            list(1, listmax) = cnt;
-            ++listmax;
-        }
-    }
-    windowshadow = 1;
-label_1857_internal:
-    cs_bk = -1;
-    pagemax = (listmax - 1) / pagesize;
-    if (page < 0)
-    {
-        page = pagemax;
-    }
-    else if (page > pagemax)
-    {
-        page = 0;
-    }
-    s(0) = lang(u8"生産品の選択"s, u8"Production"s);
-    s(1) = strhint2 + strhint3b;
-    display_window((windoww - 640) / 2 + inf_screenx, winposy(448), 640, 448);
-    display_topic(lang(u8"生産品"s, u8"Product"s), wx + 28, wy + 36);
-    display_topic(lang(u8"説明"s, u8"Detail"s), wx + 296, wy + 36);
-    display_topic(lang(u8"詳細"s, u8"Requirement"s), wx + 28, wy + 258);
-    display_topic(lang(u8"必要素材"s, u8"Material"s), wx + 28, wy + 304);
-    keyrange = 0;
-    for (int cnt = 0, cnt_end = (pagesize); cnt < cnt_end; ++cnt)
-    {
-        p = pagesize * page + cnt;
-        if (p >= listmax)
-        {
-            break;
-        }
-        key_list(cnt) = key_select(cnt);
-        ++keyrange;
-        if (cnt % 2 == 0)
-        {
-            pos(wx + 70, wy + 66 + cnt * 19);
-            gfini(540, 18);
-            gfdec2(12, 14, 16);
-        }
-        display_key(wx + 58, wy + 66 + cnt * 19 - 2, cnt);
-    }
-    cs_listbk();
-    f = 0;
-    for (int cnt = 0, cnt_end = (pagesize); cnt < cnt_end; ++cnt)
-    {
-        p = pagesize * page + cnt;
-        if (p >= listmax)
-        {
-            break;
-        }
-        i(0) = list(0, p);
-        i(1) = list(1, p);
-        if (cs == cnt)
-        {
-            matid = i;
-            if (invctrl == 0)
-            {
-                get_required_craft_materials();
-                font(13 - en * 2);
-                s = lang(u8"必要スキル: "s, u8"Skill needed: "s);
-                if (matval == 178)
-                {
-                    s += lang(u8"錬金術"s, u8"Alchemy"s);
-                }
-                if (matval == 176)
-                {
-                    s += lang(u8"大工"s, u8"Carpentory"s);
-                }
-                if (matval == 179)
-                {
-                    s += lang(u8"宝石細工"s, u8"Jewelling"s);
-                }
-                if (matval == 177)
-                {
-                    s += lang(u8"裁縫"s, u8"Tailoring"s);
-                }
-                s += u8" "s + matval(1) + u8"("s + sdata(matval, 0) + u8")"s;
-                if (matval(1) <= sdata(matval, 0))
-                {
-                    color(30, 30, 200);
-                }
-                else
-                {
-                    color(200, 30, 30);
-                }
-                pos(wx + 37, wy + 288);
-                mes(s + u8" "s);
-                color(0, 0, 0);
-            }
-            for (int cnt = 0; cnt < 6; ++cnt)
-            {
-                int j0 = matneed(cnt * 2);
-                int j1 = matneed(cnt * 2 + 1);
-                if (j0 == -1)
-                {
-                    break;
-                }
-                s = matname(j0) + lang(u8" ? "s, u8" x "s) + j1 + u8"("s
-                    + mat(j0) + u8")"s;
-                if (mat(j0) >= j1)
-                {
-                    color(30, 30, 200);
-                }
-                else
-                {
-                    color(200, 30, 30);
-                }
-                pos(wx + 37 + cnt % 3 * 192, wy + 334 + cnt / 3 * 16);
-                mes(s);
-                color(0, 0, 0);
-            }
-            f = 1;
-        }
-        s = ioriginalnameref(i);
-        p(1) = ipicref(i);
-        prepare_item_image(p(1), 0);
-        s(1) = lang(u8"アイテム["s + s + u8"]"s, u8"Make ["s + s + u8"]"s);
-        font(14 - en * 2);
-        if (elona::stoi(listn(0, p)) == -1)
-        {
-            p(2) = 3;
-        }
-        else
-        {
-            p(2) = 0;
-        }
-        cs_list(cs == cnt, cnven(s), wx + 86, wy + 66 + cnt * 19 - 1, 0, p(2));
-        pos(wx + 308, wy + 66 + cnt * 19 + 2);
-        mes(s(1));
-        pos(wx + 37, wy + 69 + cnt * 19 + 2);
-        gmode(2, inf_tiles, inf_tiles);
-        grotate(1, 0, 960, 0, chipi(2, p(1)), chipi(3, p(1)));
-    }
-    if (keyrange != 0)
-    {
-        cs_bk = cs;
-    }
-    if (f == 1 || listmax == 0)
-    {
-        redraw();
-    }
-    await(config::instance().wait1);
-    key_check();
-    cursor_check();
-    ELONA_GET_SELECTED_ITEM(p, 0);
-    if (p != -1)
-    {
-        matid = p;
-        get_required_craft_materials();
-        s = ioriginalnameref(matid);
-        if (prodcheck() == -1)
-        {
-            snd(27);
-            txt(lang(
-                u8"生産の条件を満たしてない。"s,
-                u8"You don't meet requirements to create the item."s));
-            goto label_1857_internal;
-        }
-        if (!inv_getspace(0))
-        {
-            snd(27);
-            txt(lang(
-                u8"バックパックが一杯だ。"s, u8"Your inventory is full."s));
-            goto label_1857_internal;
-        }
-        matuse = 0;
-        for (int cnt = 0; cnt < 6; ++cnt)
-        {
-            int j0 = matneed(cnt * 2);
-            int j1 = matneed(cnt * 2 + 1);
-            if (j0 == -1)
-            {
-                break;
-            }
-            mat(j0) -= j1;
-            matuse += j1;
-        }
-        snd(58);
-        fixlv = 2;
-        if (rnd(200 + matval(1) * 2) < sdata(matval, 0) + 20)
-        {
-            fixlv = 4;
-        }
-        if (rnd(100 + matval(1) * 2) < sdata(matval, 0) + 20)
-        {
-            fixlv = 3;
-        }
-        flt(calcobjlv(sdata(matval, 0)), calcfixlv(fixlv));
-        itemcreate(0, p, -1, -1, 0);
-        txt(lang(
-            itemname(ci, 1) + u8"を製造した。"s,
-            u8"You producted "s + itemname(ci, 1) + u8"."s));
-        r2 = matuse;
-        label_1472(matval);
-        refresh_character(0);
-        render_hud();
-        page_save();
-        goto label_18551_internal;
-    }
-    if (key == key_pageup)
-    {
-        if (pagemax != 0)
-        {
-            snd(1);
-            ++page;
-            goto label_1857_internal;
-        }
-    }
-    if (key == key_pagedown)
-    {
-        if (pagemax != 0)
-        {
-            snd(1);
-            --page;
-            goto label_1857_internal;
-        }
-    }
-    if (key == key_cancel)
-    {
-        return;
-    }
-    goto label_1857_internal;
-}
-
-
 
 menu_result menu_materials()
 {
@@ -36517,1886 +35162,6 @@ void label_1901()
     return;
 }
 
-
-
-std::string rpmatname(int prm_1037)
-{
-    std::string s_at_m177;
-    int p_at_m177 = 0;
-    s_at_m177 = u8"?????"s;
-    if (rpdata(20 + prm_1037, rpid) < 9000)
-    {
-        s_at_m177 = ioriginalnameref(rpdata(20 + prm_1037, rpid));
-    }
-    else if (rpdata(20 + prm_1037, rpid) < 10000)
-    {
-        s_at_m177 = rfnameorg(1, rpdata(20 + prm_1037, rpid) - 9000);
-    }
-    else
-    {
-        s_at_m177 = fltname(rpdata(20 + prm_1037, rpid));
-    }
-    if (rpdata(40 + prm_1037, rpid) == 0)
-    {
-        return s_at_m177;
-    }
-    if (rpdata(40 + prm_1037, rpid) >= 10000)
-    {
-        p_at_m177 = rpdata((40 + prm_1037), rpid) % 10000;
-        if (p_at_m177 < 0 || p_at_m177 >= 800)
-        {
-            s_at_m177 += u8"/bugged/"s;
-            return s_at_m177;
-        }
-        s_at_m177 =
-            refchara_str(p_at_m177, 2) + lang(u8"の死体"s, u8" corpse"s);
-        return s_at_m177;
-    }
-    return s_at_m177;
-}
-
-
-
-int blendcheckext(int prm_1038, int prm_1039)
-{
-    int p_at_m178 = 0;
-    if (rpdata(40 + prm_1039, rpid) >= 10000)
-    {
-        p_at_m178 = rpdata((40 + prm_1039), rpid) % 10000;
-        if (p_at_m178 < 0 || p_at_m178 >= 800)
-        {
-            return 0;
-        }
-        if (inv[prm_1038].subname == p_at_m178)
-        {
-            return 1;
-        }
-        else
-        {
-            return 0;
-        }
-    }
-    return 0;
-}
-
-
-
-std::string rpsuccessrate(int prm_1040)
-{
-    if (prm_1040 == 100)
-    {
-        return lang(u8"もう完璧！"s, u8"Perfect!"s);
-    }
-    if (prm_1040 >= 90)
-    {
-        return lang(u8"朝飯前！"s, u8"A piece of cake!"s);
-    }
-    if (prm_1040 >= 80)
-    {
-        return lang(u8"かんたんね"s, u8"Very likely"s);
-    }
-    if (prm_1040 >= 70)
-    {
-        return lang(u8"まず大丈夫"s, u8"No problem"s);
-    }
-    if (prm_1040 >= 60)
-    {
-        return lang(u8"たぶんいける"s, u8"Probably OK"s);
-    }
-    if (prm_1040 >= 50)
-    {
-        return lang(u8"ちょっと不安"s, u8"Maybe"s);
-    }
-    if (prm_1040 >= 40)
-    {
-        return lang(u8"だめかも"s, u8"Bad"s);
-    }
-    if (prm_1040 >= 30)
-    {
-        return lang(u8"やばい"s, u8"Very bad"s);
-    }
-    if (prm_1040 >= 20)
-    {
-        return lang(u8"まず無理"s, u8"Almost impossible"s);
-    }
-    return lang(u8"絶対ムリ！"s, u8"Impossible!"s);
-}
-
-
-
-int rpdiff(int, int prm_1042, int prm_1043)
-{
-    int p1_at_m180 = 0;
-    int f_at_m180 = 0;
-    int f2_at_m180 = 0;
-    int i_at_m180 = 0;
-    int d_at_m180 = 0;
-    int p_at_m180 = 0;
-    p1_at_m180 = 80;
-    f_at_m180 = 100;
-    if (prm_1042 > 0)
-    {
-        for (int cnt = 0, cnt_end = (prm_1042); cnt < cnt_end; ++cnt)
-        {
-            f2_at_m180 = 0;
-            if (rpref(10 + cnt * 2) == -1)
-            {
-                break;
-            }
-            i_at_m180 = rpref(10 + cnt * 2);
-            if (inv[i_at_m180].curse_state == curse_state_t::blessed)
-            {
-                f2_at_m180 -= 10;
-            }
-            if (is_cursed(inv[i_at_m180].curse_state))
-            {
-                f2_at_m180 += 20;
-            }
-            f_at_m180 += f2_at_m180;
-            if (prm_1043 == cnt)
-            {
-                while (1)
-                {
-                    if (f2_at_m180 < 0)
-                    {
-                        txtef(2);
-                        txt(lang(
-                            u8"調合の成功率が上がった。"s,
-                            u8"The success rate goes up."s));
-                        break;
-                    }
-                    if (f2_at_m180 > 0)
-                    {
-                        txtef(3);
-                        txt(lang(
-                            u8"調合の成功率が下がった。"s,
-                            u8"The success rate goes down."s));
-                        break;
-                    }
-                    break;
-                }
-            }
-        }
-    }
-    for (int cnt = 0; cnt < 5; ++cnt)
-    {
-        if (rpdata(10 + cnt * 2, rpid) == 0)
-        {
-            break;
-        }
-        if (sdata(rpdata(10 + cnt * 2, rpid), 0) <= 0)
-        {
-            p1_at_m180 -= 125;
-            continue;
-        }
-        d_at_m180 = rpdata(11 + cnt * 2, rpid);
-        if (prm_1042 > 0)
-        {
-            d_at_m180 = d_at_m180 * f_at_m180 / 100;
-            if (d_at_m180 < 1)
-            {
-                d_at_m180 = 1;
-            }
-        }
-        p_at_m180 =
-            (d_at_m180 * 200 / sdata(rpdata((10 + cnt * 2), rpid), 0) - 200)
-            * -1;
-        if (p_at_m180 > 0)
-        {
-            p_at_m180 /= 5;
-        }
-        if (p_at_m180 < -125)
-        {
-            p_at_m180 = -125;
-        }
-        p1_at_m180 += p_at_m180;
-    }
-    if (p1_at_m180 < 25)
-    {
-        p1_at_m180 = 0;
-    }
-    if (p1_at_m180 > 100)
-    {
-        p1_at_m180 = 100;
-    }
-    return p1_at_m180;
-}
-
-
-
-int blendcheckmat(int prm_1044)
-{
-    int f_at_m181 = 0;
-    int step_at_m181 = 0;
-    int id_at_m181 = 0;
-    int rp_at_m181 = 0;
-    int o_at_m181 = 0;
-    rpid = prm_1044;
-    for (int cnt = 0; cnt < 10; ++cnt)
-    {
-        if (rpdata(20 + cnt, rpid) == 0)
-        {
-            break;
-        }
-        f_at_m181 = 0;
-        step_at_m181 = cnt;
-        id_at_m181 = rpdata(20 + cnt, rpid);
-        rp_at_m181 = cnt;
-        for (int cnt = 0; cnt < 2; ++cnt)
-        {
-            if (cnt == 0)
-            {
-                o_at_m181 = -1;
-            }
-            if (cnt == 1)
-            {
-                o_at_m181 = 0;
-            }
-            for (const auto& cnt : items(o_at_m181))
-            {
-                if (inv[cnt].number <= 0)
-                {
-                    continue;
-                }
-                if ((rpdata(2, rpid) <= 0 || step_at_m181 != 0)
-                    && inv[cnt].own_state > 0)
-                {
-                    continue;
-                }
-                if (o_at_m181 == -1)
-                {
-                    if (dist(
-                            inv[cnt].position.x,
-                            inv[cnt].position.y,
-                            cdata[0].position.x,
-                            cdata[0].position.y)
-                        > 4)
-                    {
-                        continue;
-                    }
-                }
-                if (rpdata(40 + rp_at_m181, rpid))
-                {
-                    int stat = blendcheckext(cnt, rp_at_m181);
-                    if (stat == 0)
-                    {
-                        continue;
-                    }
-                }
-                if (id_at_m181 < 9000)
-                {
-                    if (inv[cnt].id == id_at_m181)
-                    {
-                        f_at_m181 = 1;
-                        break;
-                    }
-                    continue;
-                }
-                if (id_at_m181 < 10000)
-                {
-                    if (instr(
-                            the_item_db[inv[cnt].id]->rffilter,
-                            0,
-                            u8"/"s + rfnameorg(0, (id_at_m181 - 9000)) + u8"/"s)
-                            != -1
-                        || id_at_m181 == 9004)
-                    {
-                        f_at_m181 = 1;
-                        break;
-                    }
-                    continue;
-                }
-                if (the_item_db[inv[cnt].id]->category == id_at_m181)
-                {
-                    f_at_m181 = 1;
-                    break;
-                }
-            }
-            if (f_at_m181 == 1)
-            {
-                break;
-            }
-        }
-        if (f_at_m181 == 0)
-        {
-            break;
-        }
-    }
-    return f_at_m181;
-}
-
-
-
-int blendmatnum(int prm_1045, int prm_1046)
-{
-    int m_at_m182 = 0;
-    int o_at_m182 = 0;
-    m_at_m182 = 0;
-    for (int cnt = 0; cnt < 2; ++cnt)
-    {
-        if (cnt == 0)
-        {
-            o_at_m182 = -1;
-        }
-        if (cnt == 1)
-        {
-            o_at_m182 = 0;
-        }
-        for (const auto& cnt : items(o_at_m182))
-        {
-            if (inv[cnt].number <= 0)
-            {
-                continue;
-            }
-            if ((rpdata(2, rpid) <= 0 || prm_1046 != 0)
-                && inv[cnt].own_state > 0)
-            {
-                continue;
-            }
-            if (o_at_m182 == -1)
-            {
-                if (dist(
-                        inv[cnt].position.x,
-                        inv[cnt].position.y,
-                        cdata[0].position.x,
-                        cdata[0].position.y)
-                    > 4)
-                {
-                    continue;
-                }
-            }
-            if (rpdata(40 + prm_1046, rpid))
-            {
-                int stat = blendcheckext(cnt, prm_1046);
-                if (stat == 0)
-                {
-                    continue;
-                }
-            }
-            if (prm_1045 < 9000)
-            {
-                if (inv[cnt].id == prm_1045)
-                {
-                    m_at_m182 += inv[cnt].number;
-                }
-                continue;
-            }
-            if (prm_1045 < 10000)
-            {
-                if (instr(
-                        the_item_db[inv[cnt].id]->rffilter,
-                        0,
-                        u8"/"s + rfnameorg(0, (prm_1045 - 9000)) + u8"/"s)
-                        != -1
-                    || prm_1045 == 9004)
-                {
-                    m_at_m182 += inv[cnt].number;
-                }
-                continue;
-            }
-            if (the_item_db[inv[cnt].id]->category == prm_1045)
-            {
-                m_at_m182 += inv[cnt].number;
-                continue;
-            }
-        }
-    }
-    return m_at_m182;
-}
-
-
-
-int blendlist(elona_vector2<int>& prm_1047, int prm_1048)
-{
-    int id_at_m183 = 0;
-    int m_at_m183 = 0;
-    int o_at_m183 = 0;
-    int reftype_at_m183 = 0;
-    int f_at_m183 = 0;
-    id_at_m183 = rpdata(20 + prm_1048, rpid);
-    m_at_m183 = 0;
-    for (int cnt = 0; cnt < 2; ++cnt)
-    {
-        if (cnt == 0)
-        {
-            o_at_m183 = -1;
-        }
-        if (cnt == 1)
-        {
-            o_at_m183 = 0;
-        }
-        for (const auto& cnt : items(o_at_m183))
-        {
-            if (m_at_m183 >= 500)
-            {
-                break;
-            }
-            if (inv[cnt].number <= 0)
-            {
-                continue;
-            }
-            if ((rpdata(2, rpid) <= 0 || step != 0) && inv[cnt].own_state > 0)
-            {
-                continue;
-            }
-            if (o_at_m183 == -1)
-            {
-                if (dist(
-                        inv[cnt].position.x,
-                        inv[cnt].position.y,
-                        cdata[0].position.x,
-                        cdata[0].position.y)
-                    > 4)
-                {
-                    continue;
-                }
-            }
-            reftype_at_m183 = the_item_db[inv[cnt].id]->category;
-            if (rpdata(40 + prm_1048, rpid))
-            {
-                int stat = blendcheckext(cnt, prm_1048);
-                if (stat == 0)
-                {
-                    continue;
-                }
-            }
-            if (id_at_m183 < 9000)
-            {
-                if (inv[cnt].id != id_at_m183)
-                {
-                    continue;
-                }
-            }
-            else if (id_at_m183 < 10000)
-            {
-                if (instr(
-                        the_item_db[inv[cnt].id]->rffilter,
-                        0,
-                        u8"/"s + rfnameorg(0, (id_at_m183 - 9000)) + u8"/"s)
-                        == -1
-                    && id_at_m183 != 9004)
-                {
-                    continue;
-                }
-            }
-            else if (reftype_at_m183 != id_at_m183)
-            {
-                continue;
-            }
-            if (step > 0)
-            {
-                f_at_m183 = cnt;
-                for (int cnt = 0, cnt_end = (step); cnt < cnt_end; ++cnt)
-                {
-                    if (rpref(10 + cnt * 2) == f_at_m183)
-                    {
-                        f_at_m183 = -999;
-                        break;
-                    }
-                }
-                if (f_at_m183 == -999)
-                {
-                    continue;
-                }
-            }
-            prm_1047(0, m_at_m183) = cnt;
-            prm_1047(1, m_at_m183) = reftype_at_m183 * 1000 + inv[cnt].id;
-            ++m_at_m183;
-        }
-    }
-    return m_at_m183;
-}
-
-
-
-void window_recipe2(int val0)
-{
-    int x_at_m183 = 0;
-    int w_at_m183 = 0;
-    int dx_at_m183 = 0;
-    int dy_at_m183 = 0;
-    std::string s_at_m183;
-    int p_at_m183 = 0;
-    x_at_m183 = wx + ww;
-    w_at_m183 = 400;
-    gmode(2);
-    pos(x_at_m183 + w_at_m183 - 520, 0);
-    gcopy(3, 960, 288, 480, 68);
-    dx_at_m183 = x_at_m183 + w_at_m183 - 500;
-    dy_at_m183 = 10;
-    font(15 - en * 2, snail::font_t::style_t::bold);
-    s_at_m183 = ""s + rpsuccessrate(rpdiff(rpid, step, -1));
-    pos(dx_at_m183 + 140, dy_at_m183);
-    color(30, 30, 30);
-    bmes(lang(u8"成功率: "s, u8"Success Rate: "s) + s_at_m183, 235, 235, 235);
-    color(0, 0, 0);
-    p_at_m183 = rpdata(1, rpid);
-    if (rpmode)
-    {
-        if (p_at_m183 < 10000)
-        {
-            p_at_m183 = p_at_m183 * val0;
-        }
-        else
-        {
-            p_at_m183 = p_at_m183 % 10000;
-        }
-        p_at_m183 += rpdata(1, rpid) / 10000 * val0 * 10000;
-    }
-    s_at_m183 = ""s + p_at_m183 % 10000 + lang(u8"ターン"s, u8" turns"s);
-    if (p_at_m183 >= 10000)
-    {
-        s_at_m183 += lang(u8"と"s, u8" and "s) + p_at_m183 / 10000
-            + lang(u8"時間"s, u8" hours"s);
-    }
-    pos(dx_at_m183 + 140, dy_at_m183 + 20);
-    color(40, 40, 40);
-    bmes(lang(u8"必要時間: "s, u8"Time: "s) + s_at_m183, 235, 235, 235);
-    color(0, 0, 0);
-    return;
-}
-
-
-
-void window_recipe_(
-    int prm_1050,
-    int prm_1051,
-    int prm_1052,
-    int prm_1053,
-    int prm_1054)
-{
-    elona_vector1<std::string> s_at_m184;
-    int xfix2_at_m184 = 0;
-    int dx_at_m184 = 0;
-    int dy_at_m184 = 0;
-    int i_at_m184 = 0;
-    int p_at_m184 = 0;
-    int cnt2_at_m184 = 0;
-    SDIM1(s_at_m184);
-    if (windowshadow == 1)
-    {
-        window(
-            prm_1051 + 4,
-            prm_1052 + 4,
-            prm_1053,
-            prm_1054 - prm_1054 % 8,
-            0,
-            -1);
-        pos(prm_1051 + prm_1053 - 522, 0);
-        gfini(486, 69);
-        gfdec(40, 40, 40);
-        windowshadow = 0;
-    }
-    window(prm_1051, prm_1052, prm_1053, prm_1054 - prm_1054 % 8, 0, 0);
-    window_recipe2();
-    gmode(2);
-    color(194, 170, 146);
-    line(
-        prm_1051 + 50 + 0,
-        prm_1052 + prm_1054 - 48 - prm_1054 % 8,
-        prm_1051 + prm_1053 - 40,
-        prm_1052 + prm_1054 - 48 - prm_1054 % 8);
-    color(0, 0, 0);
-    color(234, 220, 188);
-    line(
-        prm_1051 + 50 + 0,
-        prm_1052 + prm_1054 - 49 - prm_1054 % 8,
-        prm_1051 + prm_1053 - 40,
-        prm_1052 + prm_1054 - 49 - prm_1054 % 8);
-    color(0, 0, 0);
-    s_at_m184(0) = u8"Page."s + (rppage + 1) + u8"/"s + (rppage(1) + 1);
-    s_at_m184(1) = ""s + key_prev + u8","s + key_next + ""s
-        + lang(u8"[ページ切替]  "s, u8"[Page]  "s);
-    if (step == -1)
-    {
-        s_at_m184(1) += strhint3;
-    }
-    else
-    {
-        s_at_m184(1) += strhint3b;
-    }
-    font(12 + sizefix - en * 2);
-    pos(prm_1051 + 25 + 0, prm_1052 + prm_1054 - 43 - prm_1054 % 8);
-    mes(s_at_m184(1));
-    font(12 + sizefix - en * 2, snail::font_t::style_t::bold);
-    pos(prm_1051 + prm_1053 - strlen_u(s_at_m184) * 7 - 40 - xfix2_at_m184,
-        prm_1052 + prm_1054 - 65 - prm_1054 % 8);
-    mes(s_at_m184);
-    dx_at_m184 = prm_1051 + 35;
-    dy_at_m184 = y + 48;
-    font(12 - en * 2, snail::font_t::style_t::bold);
-    pos(dx_at_m184 - 10, dy_at_m184);
-    mes(lang(u8"調合の手順"s, u8"Blending Procedure"s));
-    dy_at_m184 = dy_at_m184 + 18;
-    font(13 - en * 2);
-    i_at_m184 = 1;
-    pos(dx_at_m184 - 10, dy_at_m184 - 2);
-    gfini(prm_1053 - 60, 17);
-    if (step == i_at_m184 - 2)
-    {
-        gfdec2(10, 20, 30);
-    }
-    else if (step > i_at_m184 - 2)
-    {
-        gfdec2(25, 30, 35);
-    }
-    if (step == -1)
-    {
-        pos(dx_at_m184, dy_at_m184);
-        mes(""s + i_at_m184 + u8"."s
-            + lang(u8"レシピを選ぶ"s, u8"Choose a recipe"s));
-    }
-    else
-    {
-        pos(dx_at_m184, dy_at_m184);
-        mes(""s + i_at_m184 + u8"."s
-            + lang(
-                  (rpname(rpid) + u8"のレシピを選んだ"s),
-                  (u8"Chose the recipe of "s + rpname(rpid))));
-    }
-    dy_at_m184 += 17;
-    ++i_at_m184;
-    for (int cnt = 0; cnt < 10; ++cnt)
-    {
-        if (rpdata(20 + cnt, rpid) == 0)
-        {
-            break;
-        }
-        pos(dx_at_m184 - 10, dy_at_m184 - 2);
-        gfini(prm_1053 - 60, 17);
-        if (step == i_at_m184 - 2)
-        {
-            gfdec2(10, 20, 30);
-        }
-        else if (step > i_at_m184 - 2)
-        {
-            gfdec2(25, 30, 35);
-        }
-        if (step <= cnt)
-        {
-            int stat = blendmatnum(rpdata(20 + cnt, rpid), cnt);
-            p_at_m184 = stat;
-            s_at_m184 = lang(
-                            rpmatname(cnt) + u8"を加える"s,
-                            u8"Add "s + rpmatname(cnt) + ""s)
-                + u8"("s + lang(u8"所持:"s, u8"Stock:"s) + p_at_m184 + u8")"s;
-        }
-        else
-        {
-            s_at_m184 = lang(""s, u8"Selected "s)
-                + itemname(rpref((10 + cnt * 2))) + lang(u8"を選んだ"s, ""s);
-            s_at_m184 = strmid(s_at_m184, 0, 44);
-        }
-        pos(dx_at_m184, dy_at_m184);
-        mes(""s + i_at_m184 + u8"."s + s_at_m184);
-        dy_at_m184 += 17;
-        ++i_at_m184;
-    }
-    pos(wx + ww + 243, wy - 4);
-    gcopy(3, 1040, 96, 160, 70);
-    pos(dx_at_m184 - 10, dy_at_m184 - 2);
-    gfini(prm_1053 - 60, 17);
-    if (step == i_at_m184 - 2)
-    {
-        gfdec2(10, 20, 30);
-    }
-    else if (step > i_at_m184 - 2)
-    {
-        gfdec2(25, 30, 35);
-    }
-    pos(dx_at_m184, dy_at_m184);
-    mes(""s + i_at_m184 + u8"."s
-        + lang(u8"調合を始める！"s, u8"Start blending!"s));
-    dy_at_m184 += 30;
-    if (rppage == 0)
-    {
-        font(12 - en * 2, snail::font_t::style_t::bold);
-        pos(dx_at_m184 - 10, dy_at_m184);
-        mes(lang(
-            rpname(rpid) + u8"のレシピ"s, u8"The recipe of "s + rpname(rpid)));
-        dy_at_m184 += 20;
-        pos(dx_at_m184 - 10, dy_at_m184);
-        mes(lang(u8"必要なスキル:"s, u8"Required Skills:"s));
-        dy_at_m184 = dy_at_m184 + 18;
-        font(13 - en * 2);
-        for (int cnt = 0; cnt < 5; ++cnt)
-        {
-            if (rpdata(10 + cnt * 2, rpid) == 0)
-            {
-                break;
-            }
-            if (rpdata(11 + cnt * 2, rpid)
-                > sdata(rpdata(10 + cnt * 2, rpid), 0))
-            {
-                color(150, 0, 0);
-            }
-            else
-            {
-                color(0, 120, 0);
-            }
-            pos(dx_at_m184 + cnt % 2 * 140, dy_at_m184 + cnt / 2 * 17);
-            mes(i18n::_(
-                    u8"ability",
-                    std::to_string(rpdata(10 + cnt * 2, rpid)),
-                    u8"name")
-                + u8"  "s + rpdata((11 + cnt * 2), rpid) + u8"("s
-                + sdata(rpdata((10 + cnt * 2), rpid), 0) + u8")"s);
-            color(0, 0, 0);
-        }
-        dy_at_m184 += 50;
-        font(12 - en * 2, snail::font_t::style_t::bold);
-        pos(dx_at_m184 - 10, dy_at_m184);
-        mes(lang(u8"必要な機材:"s, u8"Required equipment:"s));
-        return;
-    }
-    if (prm_1050 == -1)
-    {
-        return;
-    }
-    font(12 - en * 2, snail::font_t::style_t::bold);
-    pos(dx_at_m184 - 10, dy_at_m184);
-    mes(itemname(prm_1050));
-    dy_at_m184 += 20;
-    font(13 - en * 2);
-    if (inv[prm_1050].identification_state
-        <= identification_state_t::partly_identified)
-    {
-        pos(dx_at_m184, dy_at_m184);
-        mes(lang(
-            u8"このアイテムは鑑定されていない。"s,
-            u8"You haven't identified it yet."s));
-        dy_at_m184 += 16;
-        return;
-    }
-    getinheritance(prm_1050, inhlist_at_m184, p_at_m184);
-    if (p_at_m184 > 0)
-    {
-        for (int cnt = 0, cnt_end = (p_at_m184); cnt < cnt_end; ++cnt)
-        {
-            cnt2_at_m184 = inhlist_at_m184(cnt);
-            if (inv[prm_1050].enchantments[cnt2_at_m184].id == 0)
-            {
-                break;
-            }
-            get_enchantment_description(
-                inv[prm_1050].enchantments[cnt2_at_m184].id,
-                inv[prm_1050].enchantments[cnt2_at_m184].power,
-                the_item_db[inv[prm_1050].id]->category);
-            color(0, 0, 100);
-            if (inv[prm_1050].enchantments[cnt2_at_m184].power < 0)
-            {
-                color(180, 0, 0);
-            }
-            pos(dx_at_m184, dy_at_m184);
-            mes(cnven(s));
-            color(0, 0, 0);
-            dy_at_m184 += 16;
-        }
-    }
-    else
-    {
-        pos(dx_at_m184, dy_at_m184);
-        mes(lang(u8"継承効果なし"s, u8"No inheritance effects"s));
-        dy_at_m184 += 16;
-        ++p_at_m184;
-    }
-    return;
-}
-
-
-
-void clear_rprefmat()
-{
-    for (int cnt = 0; cnt < 10; ++cnt)
-    {
-        rpref(10 + cnt * 2) = -1;
-    }
-    return;
-}
-
-
-
-void initialize_recipememory()
-{
-    for (int cnt = 0; cnt < 8; ++cnt)
-    {
-        recipememory(200 + cnt) = 1;
-    }
-    return;
-}
-
-
-
-void initialize_recipe()
-{
-    DIM3(rpdata, 100, 1200);
-    SDIM4(rfnameorg, 20, 2, 6);
-    SDIM3(rpdatan, 40, 1200);
-    DIM2(rpref, 100);
-    rpsourcelist(0) = 0;
-    rpsourcelist(1) = 898;
-    rpsourcelist(2) = 1156;
-    rpsourcelist(3) = 1187;
-    rpsourcelist(4) = 209;
-    rpsourcelist(5) = 210;
-    rpid = 200;
-    rpdatan(rpid) = lang(u8"媚薬混入食品"s, u8"love food"s);
-    rpdata(0, rpid) = 10000;
-    rpdata(1, rpid) = 10;
-    rpdata(2, rpid) = 1;
-    rpdata(10, rpid) = 184;
-    rpdata(11, rpid) = 8;
-    rpdata(12, rpid) = 178;
-    rpdata(13, rpid) = 3;
-    rpdata(20, rpid) = 57000;
-    rpdata(21, rpid) = 620;
-    rpid = 201;
-    rpdatan(rpid) = lang(u8"染色"s, u8"dyeing"s);
-    rpdata(0, rpid) = 10001;
-    rpdata(1, rpid) = 4;
-    rpdata(2, rpid) = 1;
-    rpdata(10, rpid) = 12;
-    rpdata(11, rpid) = 6;
-    rpdata(20, rpid) = 9004;
-    rpdata(21, rpid) = 519;
-    rpid = 202;
-    rpdatan(rpid) = lang(u8"特製毒入り食品"s, u8"poisoned food"s);
-    rpdata(0, rpid) = 10002;
-    rpdata(1, rpid) = 7;
-    rpdata(2, rpid) = 1;
-    rpdata(10, rpid) = 184;
-    rpdata(11, rpid) = 3;
-    rpdata(20, rpid) = 57000;
-    rpdata(21, rpid) = 262;
-    rpid = 203;
-    rpdatan(rpid) = lang(u8"耐火コーティング"s, u8"fireproof coating"s);
-    rpdata(0, rpid) = 10003;
-    rpdata(1, rpid) = 15;
-    rpdata(2, rpid) = 1;
-    rpdata(10, rpid) = 12;
-    rpdata(11, rpid) = 18;
-    rpdata(20, rpid) = 9004;
-    rpdata(21, rpid) = 736;
-    rpid = 204;
-    rpdatan(rpid) = lang(u8"耐酸コーティング"s, u8"acidproof coating"s);
-    rpdata(0, rpid) = 10004;
-    rpdata(1, rpid) = 15;
-    rpdata(2, rpid) = 1;
-    rpdata(10, rpid) = 12;
-    rpdata(11, rpid) = 10;
-    rpdata(20, rpid) = 9004;
-    rpdata(21, rpid) = 566;
-    rpid = 205;
-    rpdatan(rpid) = lang(u8"釣り餌の装着"s, u8"bait attachment"s);
-    rpdata(0, rpid) = 10005;
-    rpdata(1, rpid) = 10;
-    rpdata(2, rpid) = 1;
-    rpdata(10, rpid) = 185;
-    rpdata(11, rpid) = 2;
-    rpdata(12, rpid) = 12;
-    rpdata(13, rpid) = 10;
-    rpdata(20, rpid) = 342;
-    rpdata(21, rpid) = 617;
-    rpid = 206;
-    rpdatan(rpid) = lang(u8"アイテムの祝福"s, u8"blessed item"s);
-    rpdata(0, rpid) = 10006;
-    rpdata(1, rpid) = 5;
-    rpdata(2, rpid) = 1;
-    rpdata(10, rpid) = 12;
-    rpdata(11, rpid) = 5;
-    rpdata(20, rpid) = 9004;
-    rpdata(21, rpid) = 516;
-    rpid = 207;
-    rpdatan(rpid) = lang(u8"井戸水の回復"s, u8"well refill"s);
-    rpdata(0, rpid) = 10007;
-    rpdata(1, rpid) = 3;
-    rpdata(2, rpid) = 2;
-    rpdata(10, rpid) = 12;
-    rpdata(11, rpid) = 3;
-    rpdata(20, rpid) = 60001;
-    rpdata(21, rpid) = 52000;
-    rpid = 208;
-    rpdatan(rpid) = lang(u8"天然ポーション"s, u8"natural potion"s);
-    rpdata(0, rpid) = 10008;
-    rpdata(1, rpid) = 16;
-    rpdata(2, rpid) = 2;
-    rpdata(10, rpid) = 12;
-    rpdata(11, rpid) = 24;
-    rpdata(20, rpid) = 60001;
-    rpdata(21, rpid) = 601;
-    rpid = 209;
-    rpdatan(rpid) = lang(u8"2種アーティファクト合成"s, u8"2 artifacts fusion"s);
-    rpdata(0, rpid) = 10009;
-    rpdata(1, rpid) = 16;
-    rpdata(2, rpid) = 2;
-    rpdata(10, rpid) = 178;
-    rpdata(11, rpid) = 999;
-    rpdata(20, rpid) = 9004;
-    rpdata(21, rpid) = 9004;
-    rpid = 210;
-    rpdatan(rpid) = lang(u8"3種アーティファクト合成"s, u8"3 artifacts fusion"s);
-    rpdata(0, rpid) = 10009;
-    rpdata(1, rpid) = 16;
-    rpdata(2, rpid) = 2;
-    rpdata(10, rpid) = 178;
-    rpdata(11, rpid) = 999;
-    rpdata(20, rpid) = 9004;
-    rpdata(21, rpid) = 9004;
-    rpdata(22, rpid) = 9004;
-    rpid = 898;
-    rpdata(0, rpid) = 498;
-    rpdata(1, rpid) = 150060;
-    rpdata(10, rpid) = 175;
-    rpdata(11, rpid) = 4;
-    rpdata(12, rpid) = 457;
-    rpdata(13, rpid) = 20;
-    rpdata(14, rpid) = 12;
-    rpdata(15, rpid) = 8;
-    rpdata(30, rpid) = 0;
-    rpdata(20, rpid) = 260;
-    rpdata(21, rpid) = 9001;
-    rpid = 1156;
-    rpdata(0, rpid) = 756;
-    rpdata(1, rpid) = 20020;
-    rpdata(10, rpid) = 184;
-    rpdata(11, rpid) = 4;
-    rpdata(12, rpid) = 179;
-    rpdata(13, rpid) = 20;
-    rpdata(14, rpid) = 157;
-    rpdata(15, rpid) = 8;
-    rpdata(30, rpid) = 0;
-    rpdata(20, rpid) = 260;
-    ++rpid;
-    rpid = 1187;
-    rpdata(0, rpid) = 787;
-    rpdata(1, rpid) = 30;
-    rpdata(10, rpid) = 184;
-    rpdata(11, rpid) = 10;
-    rpdata(12, rpid) = 178;
-    rpdata(13, rpid) = 5;
-    rpdata(30, rpid) = 0;
-    rpdata(20, rpid) = 9004;
-    rpid = 1191;
-    rpdata(0, rpid) = 791;
-    rpdata(1, rpid) = 160045;
-    rpdata(10, rpid) = 184;
-    rpdata(11, rpid) = 10;
-    rpdata(12, rpid) = 178;
-    rpdata(13, rpid) = 65;
-    rpdata(14, rpid) = 10;
-    rpdata(15, rpid) = 105;
-    rpdata(30, rpid) = 0;
-    rpdata(20, rpid) = 10000;
-    rpdata(21, rpid) = 748;
-    rpdata(22, rpid) = 716;
-    rpid = 1192;
-    rpdata(0, rpid) = 792;
-    rpdata(1, rpid) = 10024;
-    rpdata(10, rpid) = 184;
-    rpdata(11, rpid) = 20;
-    rpdata(12, rpid) = 0;
-    rpdata(13, rpid) = 10;
-    rpdata(14, rpid) = 151;
-    rpdata(15, rpid) = 5;
-    rpdata(30, rpid) = 0;
-    rpdata(20, rpid) = 204;
-    rpdata(21, rpid) = 9005;
-    rpdata(40, rpid) = 10003;
-    rpdata(50, rpid) = 10017;
-    rpdata(51, rpid) = 500;
-    rpdata(52, rpid) = 60017;
-    rpdata(53, rpid) = 500;
-    rpdata(54, rpid) = 10010;
-    rpdata(55, rpid) = 100;
-    rpdata(56, rpid) = 60010;
-    rpdata(57, rpid) = 100;
-    rpdata(58, rpid) = 10016;
-    rpdata(59, rpid) = 300;
-    rpdata(60, rpid) = 60016;
-    rpdata(61, rpid) = 300;
-    rpdata(62, rpid) = 10011;
-    rpdata(63, rpid) = 200;
-    rpdata(64, rpid) = 60011;
-    rpdata(65, rpid) = 200;
-    rfnameorg(0, 1) = u8"flavor"s;
-    rfnameorg(1, 1) = lang(u8"適当な調味料"s, u8"suitable flavoring"s);
-    rfnameorg(0, 2) = u8"ore"s;
-    rfnameorg(1, 2) = lang(u8"適当な鉱石"s, u8"any ore"s);
-    rfnameorg(0, 3) = u8"wood"s;
-    rfnameorg(1, 3) = lang(u8"木材を含む何か"s, u8"something made of wood"s);
-    rfnameorg(0, 5) = u8"fish"s;
-    rfnameorg(1, 5) = lang(u8"魚介類"s, u8"fish"s);
-    rfnameorg(0, 4) = "";
-    rfnameorg(1, 4) = lang(u8"何か物体"s, u8"any item"s);
-    return;
-}
-
-
-
-turn_result_t crafting_menu()
-{
-    elona_vector1<int> blendchecklist;
-    step = -1;
-    rpid = 0;
-    gsel(3);
-    pos(960, 96);
-    picload(filesystem::dir::graphic() / u8"deco_blend.bmp", 1);
-    gsel(0);
-    clear_rprefmat();
-label_1923:
-    if (step != -1)
-    {
-        if (rpdata(20 + step, rpid) == 0)
-        {
-            rppage = 0;
-            window_recipe(list2, -1, wx + ww, wy, 400, wh);
-            txtnew();
-            txt(lang(
-                u8"幾つ作る？"s, u8"How many items do you want to create?"s));
-            ELONA_APPEND_PROMPT(
-                lang(u8"調合を始める"s, u8"Start blending"s),
-                u8"a"s,
-                ""s + promptmax);
-            ELONA_APPEND_PROMPT(
-                lang(u8"前に戻る"s, u8"Go back"s), u8"b"s, ""s + promptmax);
-            ELONA_APPEND_PROMPT(
-                lang(u8"最初から"s, u8"From the start"s),
-                u8"c"s,
-                ""s + promptmax);
-            p = 10;
-            for (int cnt = 0; cnt < 10; ++cnt)
-            {
-                if (rpref(10 + cnt * 2) == -1)
-                {
-                    break;
-                }
-                if (rpdata(2, rpid) == 2 && cnt == 0)
-                {
-                    continue;
-                }
-                if (inv[rpref(10 + cnt * 2)].number < p)
-                {
-                    p = inv[rpref(10 + cnt * 2)].number;
-                }
-            }
-            rpmode = 1;
-            rtval = show_prompt(
-                promptx, prompty, 220, show_prompt_type::with_number, p);
-            rpmode = 0;
-            if (rtval == 0)
-            {
-                rpref(1) = TODO_show_prompt_val;
-                rpref(2) = rpdata(1, rpid);
-                rpref(3) = rpdiff(rpid, step, -1);
-                label_19342();
-                return turn_result_t::turn_end;
-            }
-            if (rtval == 2)
-            {
-                step = -1;
-                goto label_1923;
-            }
-            --step;
-            goto label_1923;
-        }
-    }
-    page = 0;
-    pagesize = 16;
-    listmax = 0;
-    cs = 0;
-    cs_bk = -1;
-    cc = 0;
-    screenupdate = -1;
-    update_screen();
-    if (step == -1)
-    {
-        rppage(0) = 0;
-        rppage(1) = 0;
-        listmax = 0;
-        for (int cnt = 0; cnt < 1200; ++cnt)
-        {
-            if (rpdata(0, cnt) == 0)
-            {
-                continue;
-            }
-            if (recipememory(cnt) > 0)
-            {
-                list(0, listmax) = cnt;
-                list(1, listmax) = cnt;
-                ++listmax;
-            }
-        }
-        sort_list_by_column1();
-        windowshadow(1) = 1;
-        txtnew();
-        txt(lang(
-            u8"(製作中)だめまだ"s,
-            u8"(*) The feature is not implemented yet."s));
-        txtnew();
-        txt(lang(
-            u8"どのレシピを使う？"s, u8"Which recipe do you want to use?"s));
-        goto label_1924_internal;
-    }
-    rppage = 1;
-    {
-        int stat = blendlist(list, step);
-        listmax = stat;
-    }
-    sort_list_by_column1();
-    goto label_1927_internal;
-label_1924_internal:
-    cs_bk = -1;
-    pagemax = (listmax - 1) / pagesize;
-    if (page < 0)
-    {
-        page = pagemax;
-    }
-    else if (page > pagemax)
-    {
-        page = 0;
-    }
-    DIM2(blendchecklist, pagesize);
-    for (int cnt = 0, cnt_end = (pagesize); cnt < cnt_end; ++cnt)
-    {
-        p = pagesize * page + cnt;
-        if (p >= listmax)
-        {
-            break;
-        }
-        blendchecklist(cnt) = blendcheckmat(list(0, p));
-    }
-label_1925_internal:
-    s(0) = lang(u8"レシピの選択"s, u8"Choose a recipe"s);
-    s(1) = strhint2;
-    windowshadow = windowshadow(1);
-    display_window(
-        (windoww - 780) / 2 + inf_screenx, winposy(445), 380, 432, 74);
-    display_topic(lang(u8"レシピの名称"s, u8"Name"s), wx + 28, wy + 30);
-    s = ""s + listmax + u8" recipes"s;
-    font(12 + sizefix - en * 2, snail::font_t::style_t::bold);
-    pos(wx + 130, wy + wh - 65 - wh % 8);
-    mes(s);
-    keyrange = 0;
-    gmode(2);
-    prepare_item_image(550, 0);
-    for (int cnt = 0, cnt_end = (pagesize); cnt < cnt_end; ++cnt)
-    {
-        p = pagesize * page + cnt;
-        if (p >= listmax)
-        {
-            break;
-        }
-        key_list(cnt) = key_select(cnt);
-        ++keyrange;
-        if (cnt % 2 == 0)
-        {
-            pos(wx + 70, wy + 60 + cnt * 19);
-            gfini(ww - 100, 18);
-            gfdec2(12, 14, 16);
-        }
-        pos(wx + 37, wy + 70 + cnt * 19);
-        gmode(2, chipi(2, 550), chipi(3, 550));
-        grotate(1, 0, 960, 0, inf_tiles, inf_tiles);
-        pos(wx + 330, wy + 53 + cnt * 19);
-        if (blendchecklist(cnt) == 1)
-        {
-            gcopy(3, 336, 360, 24, 24);
-        }
-        rpid = list(0, p);
-        pos(wx + 317, wy + 60 + cnt * 19);
-        gcopy(3, 64 + (4 - rpdiff(rpid, -1, -1) / 25) * 16, 624, 16, 16);
-    }
-    font(14 - en * 2);
-    cs_listbk();
-    for (int cnt = 0, cnt_end = (pagesize); cnt < cnt_end; ++cnt)
-    {
-        p = pagesize * page + cnt;
-        if (p >= listmax)
-        {
-            break;
-        }
-        p = list(0, p);
-        rpid = p;
-        s = lang(""s, u8"Recipe of "s) + cnven(rpname(rpid))
-            + lang(u8"のレシピ"s, ""s);
-        display_key(wx + 58, wy + 60 + cnt * 19 - 2, cnt);
-        cs_list(cs == cnt, s, wx + 84, wy + 60 + cnt * 19 - 1, 0, 0, p);
-    }
-    if (cs_bk != cs)
-    {
-        rpid = list(0, pagesize * page + cs);
-        windowshadow = windowshadow(1);
-        window_recipe(list2, -1, wx + ww, wy, 400, wh);
-    }
-    if (keyrange != 0)
-    {
-        cs_bk = cs;
-    }
-    windowshadow(1) = 0;
-    pos(wx + 10, wy + wh - 100);
-    gcopy(3, 960, 96, 80, 90);
-    redraw();
-    await(config::instance().wait1);
-    key_check();
-    cursor_check();
-    ELONA_GET_SELECTED_ITEM(p, 0);
-    if (p != -1)
-    {
-        rpid = p;
-        step = 0;
-        rppage(0) = 1;
-        rppage(1) = 1;
-        rpref(0) = rpid;
-        goto label_1923;
-    }
-    if (key == key_pageup)
-    {
-        if (pagemax != 0)
-        {
-            snd(1);
-            ++page;
-            goto label_1924_internal;
-        }
-    }
-    if (key == key_pagedown)
-    {
-        if (pagemax != 0)
-        {
-            snd(1);
-            --page;
-            goto label_1924_internal;
-        }
-    }
-    if (key == key_cancel)
-    {
-        screenupdate = 0;
-        update_screen();
-        return turn_result_t::pc_turn_user_error;
-    }
-    goto label_1925_internal;
-label_1927_internal:
-    windowshadow(1) = 1;
-    txtnew();
-    txt(lang(
-        rpmatname(step) + u8"を追加しよう。"s,
-        u8"Add \""s + rpmatname(step) + u8"\"."s));
-label_1928_internal:
-    cs_bk = -1;
-    pagemax = (listmax - 1) / pagesize;
-    if (page < 0)
-    {
-        page = pagemax;
-    }
-    else if (page > pagemax)
-    {
-        page = 0;
-    }
-    if (jp)
-    {
-        s = rpmatname(step) + u8"を追加"s;
-    }
-    else
-    {
-        s = u8"Add "s + rpmatname(step);
-    }
-    s(1) = strhint2;
-    windowshadow = windowshadow(1);
-    display_window(
-        (windoww - 780) / 2 + inf_screenx, winposy(445), 380, 432, 74);
-    display_topic(lang(u8"アイテムの名称"s, u8"Name"s), wx + 28, wy + 30);
-    s = ""s + listmax + u8" items"s;
-    font(12 + sizefix - en * 2, snail::font_t::style_t::bold);
-    pos(wx + 130, wy + wh - 65 - wh % 8);
-    mes(s);
-    keyrange = 0;
-    for (int cnt = 0, cnt_end = (pagesize); cnt < cnt_end; ++cnt)
-    {
-        p = pagesize * page + cnt;
-        if (p >= listmax)
-        {
-            break;
-        }
-        key_list(cnt) = key_select(cnt);
-        ++keyrange;
-        if (cnt % 2 == 0)
-        {
-            pos(wx + 70, wy + 60 + cnt * 19);
-            gfini(ww - 100, 18);
-            gfdec2(12, 14, 16);
-        }
-    }
-    font(14 - en * 2);
-    cs_listbk();
-    for (int cnt = 0, cnt_end = (pagesize); cnt < cnt_end; ++cnt)
-    {
-        p = pagesize * page + cnt;
-        if (p >= listmax)
-        {
-            break;
-        }
-        p = list(0, p);
-        s = itemname(p, inv[p].number);
-        s = strmid(s, 0, 28);
-        if (p >= 5080)
-        {
-            s += lang(u8"(地面)"s, u8" (Ground)"s);
-        }
-        display_key(wx + 58, wy + 60 + cnt * 19 - 2, cnt);
-        p(1) = inv[p].image % 1000;
-        prepare_item_image(p(1), inv[p].color, inv[p].param1);
-        pos(wx + 37, wy + 69 + cnt * 19);
-        gmode(2, chipi(2, p(1)), chipi(3, p(1)));
-        grotate(
-            1,
-            0,
-            960,
-            0,
-            chipi(2, p(1)) * inf_tiles / chipi(3, p(1)),
-            inf_tiles);
-        if (inv[p].body_part != 0)
-        {
-            pos(wx + 46, wy + 72 + cnt * 18 - 3);
-            gcopy(3, 12, 348, 12, 12);
-        }
-        cs_list(cs == cnt, s, wx + 84, wy + 60 + cnt * 19 - 1, 0, 1, p);
-    }
-    p = list(0, pagesize * page + cs);
-    if (listmax == 0)
-    {
-        p = -1;
-    }
-    if (cs_bk != cs)
-    {
-        windowshadow = windowshadow(1);
-        window_recipe(list2, p, wx + ww, wy, 400, wh);
-    }
-    if (keyrange != 0)
-    {
-        cs_bk = cs;
-    }
-    windowshadow(1) = 0;
-    pos(wx + 10, wy + wh - 100);
-    gcopy(3, 960, 96, 80, 90);
-    redraw();
-    await(config::instance().wait1);
-    key_check();
-    cursor_check();
-    ELONA_GET_SELECTED_ITEM(p, 0);
-    if (key == key_pageup)
-    {
-        if (pagemax != 0)
-        {
-            snd(1);
-            ++page;
-            goto label_1928_internal;
-        }
-    }
-    if (key == key_pagedown)
-    {
-        if (pagemax != 0)
-        {
-            snd(1);
-            --page;
-            goto label_1928_internal;
-        }
-    }
-    if (p != -1)
-    {
-        snd(17);
-        ci = p;
-        rpref(10 + step * 2 + 0) = ci;
-        rpref(10 + step * 2 + 1) = inv[ci].id;
-        txt(lang(
-            itemname(ci) + u8"を選んだ。"s,
-            u8"You add "s + itemname(ci) + u8"."s));
-        ++step;
-        p = rpdiff(rpid, step, step - 1);
-        goto label_1923;
-    }
-    if (key == key_prev)
-    {
-        snd(1);
-        --rppage;
-        cs_bk = -1;
-        if (rppage < 0)
-        {
-            rppage = rppage(1);
-        }
-    }
-    if (key == key_next)
-    {
-        snd(1);
-        ++rppage;
-        cs_bk = -1;
-        if (rppage > rppage(1))
-        {
-            rppage = 0;
-        }
-    }
-    if (key == key_cancel)
-    {
-        --step;
-        goto label_1923;
-    }
-    goto label_1928_internal;
-}
-
-
-
-int label_1931()
-{
-    f = 1;
-    for (int cnt = 0; cnt < 10; ++cnt)
-    {
-        if (rpref(10 + cnt * 2) == -1)
-        {
-            break;
-        }
-        if (rpref(10 + cnt * 2) == -2)
-        {
-            f = 0;
-            break;
-        }
-        if (inv[rpref(10 + cnt * 2)].number <= 0
-            || inv[rpref(10 + cnt * 2)].id != rpref(11 + cnt * 2))
-        {
-            f = 0;
-            break;
-        }
-    }
-    return f;
-}
-
-
-
-int label_1932()
-{
-    p = 0;
-    for (int cnt = 0; cnt < 10; ++cnt)
-    {
-        if (rpref(10 + cnt * 2) == -1)
-        {
-            break;
-        }
-        if (rpref(10 + cnt * 2) == -2)
-        {
-            continue;
-        }
-        if ((rpdata(2, rpid) <= 0 || cnt != 0) && rpresult)
-        {
-            --inv[rpref(10 + cnt * 2)].number;
-        }
-        else if (rnd(3) == 0)
-        {
-            txt(lang(
-                itemname(rpref(10 + cnt * 2), 1) + u8"を失った。"s,
-                u8"You lose "s + itemname(rpref((10 + cnt * 2)), 1) + u8"."s));
-            --inv[rpref(10 + cnt * 2)].number;
-        }
-        if (chara_unequip(rpref(10 + cnt * 2)))
-        {
-            refresh_character(0);
-        }
-        cell_refresh(
-            inv[rpref(10 + cnt * 2)].position.x,
-            inv[rpref(10 + cnt * 2)].position.y);
-    }
-    refresh_burden_state();
-    return 1;
-}
-
-
-
-void label_1933()
-{
-    rpresult = 1;
-    if (rpdiff(rpid, -1, -1) < rnd(100))
-    {
-        rpresult = 0;
-        txtef(3);
-        txt(lang(u8"調合失敗！"s, u8"The blending attempt failed!"s));
-    }
-    else
-    {
-        if (rpdata(0, rpid) >= 10000)
-        {
-            label_1935();
-        }
-        else
-        {
-            flt();
-            nostack = 1;
-            int stat = itemcreate(
-                -1,
-                rpdata(0, rpid),
-                cdata[0].position.x,
-                cdata[0].position.y,
-                0);
-            if (stat != 0)
-            {
-                for (int cnt = 0;; ++cnt)
-                {
-                    if (rpdata(50 + cnt * 2, rpid) == 0)
-                    {
-                        break;
-                    }
-                    encadd(
-                        ci,
-                        rpdata(50 + cnt * 2, rpid),
-                        rpdata(51 + cnt * 2, rpid),
-                        0,
-                        1);
-                }
-            }
-            txtef(2);
-            txt(lang(
-                itemname(ci, 1) + u8"の作成に成功した！"s,
-                u8"You successfully create "s + itemname(ci, 1) + u8"!"s));
-            snd(17);
-        }
-        for (int cnt = 0; cnt < 5; ++cnt)
-        {
-            if (rpdata(10 + cnt * 2, rpid) == 0)
-            {
-                break;
-            }
-            skillexp(
-                rpdata(10 + cnt * 2, rpid),
-                0,
-                50 + rpdata((11 + cnt * 2), rpid) + rpref(2) / 10000 * 25,
-                2,
-                50);
-        }
-    }
-    --rpref(1);
-    label_1932();
-    return;
-}
-
-
-
-void label_19342()
-{
-label_19341_internal:
-    rpid = rpref(0);
-    if (rpid == 0)
-    {
-        rowactend(cc);
-        return;
-    }
-    if (cdata[cc].continuous_action_id == 0)
-    {
-        txtnew();
-        txt(lang(
-            name(cc) + u8"は"s + rpname(rpid) + u8"の調合をはじめた。"s,
-            name(cc) + u8" start"s + _s(cc) + u8" blending of "s + rpname(rpid)
-                + u8"."s));
-        cdata[cc].continuous_action_id = 12;
-        cdata[cc].continuous_action_turn = rpref(2) % 10000;
-        return;
-    }
-    if (cdata[cc].continuous_action_turn > 0)
-    {
-        if (rnd(30) == 0)
-        {
-            txtef(4);
-            txt(lang(u8" *こねこね* "s, u8"*pug*"s),
-                lang(u8" *トントン* "s, u8"*clank*"s));
-        }
-        return;
-    }
-    if (rpref(2) >= 10000)
-    {
-        cdata[cc].continuous_action_turn = rpref(2) / 10000;
-        for (int cnt = 0;; ++cnt)
-        {
-            mode = 12;
-            ++gdata_hour;
-            label_2736();
-            render_hud();
-            if (cnt % 5 == 0)
-            {
-                txtef(4);
-                txt(lang(u8" *こねこね* "s, u8"*pug*"s),
-                    lang(u8" *トントン* "s, u8"*clank*"s));
-            }
-            for (int cnt = 0; cnt < 20; ++cnt)
-            {
-                redraw();
-                await(30);
-            }
-            gdata_minute = 0;
-            cc = 0;
-            --cdata[cc].continuous_action_turn;
-            if (cdata[cc].continuous_action_turn <= 0)
-            {
-                int stat = label_1931();
-                if (stat == 0)
-                {
-                    txt(lang(
-                        u8"調合に必要な材料が見つからない。"s,
-                        u8"A required material cannot be found."s));
-                    break;
-                }
-                label_1933();
-                if (rpref(1) > 0)
-                {
-                    cdata[cc].continuous_action_turn = rpref(2) / 10000;
-                    cnt = 0 - 1;
-                    continue;
-                }
-                else
-                {
-                    break;
-                }
-            }
-        }
-        rowactend(cc);
-        mode = 0;
-        return;
-    }
-    int stat = label_1931();
-    if (stat == 0)
-    {
-        txt(lang(
-            u8"調合に必要な材料が見つからない。"s,
-            u8"A required material cannot be found."s));
-        rowactend(cc);
-        return;
-    }
-    label_1933();
-    if (rpref(1) > 0)
-    {
-        cdata[cc].continuous_action_id = 0;
-        goto label_19341_internal;
-    }
-    rowactend(cc);
-    return;
-}
-
-
-
-void label_1935()
-{
-    ci = rpref(10);
-    ti = rpref(12);
-    if (rpdata(2, rpid) == 2)
-    {
-        item_separate(ci);
-    }
-    else if (inv[ci].number <= 1)
-    {
-        rpref(10) = -2;
-    }
-    else
-    {
-        int stat = item_separate(ci);
-        if (rpref(10) == stat)
-        {
-            rpref(10) = -2;
-        }
-        else
-        {
-            rpref(10) = stat;
-        }
-    }
-    switch (rpdata(0, rpid))
-    {
-    case 10000:
-        ibitmod(6, ci, 1);
-        txtef(2);
-        txt(lang(
-            itemname(ci, 1) + u8"の作成に成功した！"s,
-            u8"You successfully create "s + itemname(ci, 1) + u8"!"s));
-        txt(lang(
-            u8"あなたはうしろめたさを感じた…"s,
-            u8"You kind of feel guilty..."s));
-        snd(65);
-        break;
-    case 10001:
-        inv[ci].color = inv[ti].color;
-        txtef(2);
-        txt(lang(
-            u8"あなたは"s + itemname(ci) + u8"を染めた。"s,
-            u8"You dye "s + itemname(ci) + u8"."s));
-        snd(17);
-        break;
-    case 10002:
-        ibitmod(14, ci, 1);
-        txtef(2);
-        txt(lang(
-            itemname(ci, 1) + u8"の作成に成功した！"s,
-            u8"You successfully create "s + itemname(ci, 1) + u8"!"s));
-        txt(lang(u8"あなたはにやりと口元を歪めた。"s, u8"You grin."s));
-        snd(65);
-        break;
-    case 10003:
-        txtef(2);
-        txt(lang(
-            itemname(ci) + u8"に"s + itemname(ti, 1) + u8"を塗りたくった。"s,
-            u8"You put "s + itemname(ti, 1) + u8" on "s + itemname(ci)
-                + u8"."s));
-        if (inv[ci].id == 567)
-        {
-            txt(lang(u8"いいアイデアだ！しかし…"s, u8"A good idea! But..."s));
-        }
-        else
-        {
-            ibitmod(2, ci, 1);
-            txt(lang(
-                itemname(ci) + u8"は熱から守られた。"s,
-                itemname(ci) + u8" gain"s + _s2(in) + u8" fireproof."s));
-        }
-        snd(17);
-        break;
-    case 10004:
-        txtef(2);
-        txt(lang(
-            itemname(ci) + u8"に"s + itemname(ti, 1) + u8"を塗りたくった。"s,
-            u8"You put "s + itemname(ti, 1) + u8" on "s + itemname(ci)
-                + u8"."s));
-        ibitmod(1, ci, 1);
-        txt(lang(
-            itemname(ci) + u8"は酸から守られた。"s,
-            itemname(ci) + u8" gain"s + _s2(in) + u8" acidproof."s));
-        snd(17);
-        break;
-    case 10005:
-        txtef(2);
-        txt(lang(
-            itemname(ti, 1) + u8"を"s + itemname(ci) + u8"に装着した。"s,
-            u8"You bait "s + itemname(ci) + u8" with "s + itemname(ti, 1)
-                + u8"."s));
-        if (inv[ci].param4 == inv[ti].param1)
-        {
-            inv[ci].count += rnd(10) + 15;
-        }
-        else
-        {
-            inv[ci].count = rnd(10) + 15;
-            inv[ci].param4 = inv[ti].param1;
-        }
-        snd(13);
-        break;
-    case 10006:
-        txtef(2);
-        txt(lang(
-            itemname(ti, 1) + u8"を"s + itemname(ci) + u8"に降りかけた。"s,
-            u8"You shower "s + itemname(ci) + u8" on "s + itemname(ti, 1)
-                + u8"."s));
-        if (inv[ti].curse_state == curse_state_t::blessed)
-        {
-            txtef(5);
-            txt(lang(
-                itemname(ci) + u8"は銀色に輝いた。"s,
-                itemname(ci) + u8" shine"s + _s2(inv[ci].number)
-                    + u8" silvery."s));
-            inv[ci].curse_state = curse_state_t::blessed;
-        }
-        if (is_cursed(inv[ti].curse_state))
-        {
-            txtef(8);
-            txt(lang(
-                itemname(ci) + u8"は黒いオーラに包まれた。"s,
-                itemname(ci) + u8" "s + is2(inv[ci].number)
-                    + u8" wrapped by a dark aura."s));
-            inv[ci].curse_state = curse_state_t::cursed;
-        }
-        snd(17);
-        break;
-    case 10007:
-        txt(lang(
-            itemname(ti, 1) + u8"を"s + itemname(ci) + u8"に放り込んだ。"s,
-            u8"You throw "s + itemname(ci) + u8" into "s + itemname(ti, 1)
-                + u8"."s));
-        if (inv[ti].id == 601)
-        {
-            txt(lang(
-                u8"空き瓶の割れる音がした。"s,
-                u8"You hear the sound of the empty bottle shatters."s));
-            break;
-        }
-        snd(17);
-        if (inv[ci].id == 602)
-        {
-            txt(lang(u8"井戸は汚れた。"s, u8"The holy well is polluted."s));
-            break;
-        }
-        if (inv[ci].param3 >= 20)
-        {
-            txt(lang(
-                itemname(ci) + u8"は完全に枯れている。"s,
-                itemname(ci) + u8" is completely dry."s));
-            break;
-        }
-        txtef(2);
-        txt(lang(
-            itemname(ci) + u8"は一瞬輝いた。"s,
-            itemname(ci) + u8" shines for a moment."s));
-        if (inv[ti].id == 587)
-        {
-            txt(lang(u8"しかしこんな量では… "s, u8"But the snow just melts."s));
-        }
-        else
-        {
-            inv[ci].param1 += rnd(3);
-        }
-        break;
-    case 10008:
-        if (inv[ci].param1 < -5 || inv[ci].param3 >= 20
-            || (inv[ci].id == 602 && gdata_holy_well_count <= 0))
-        {
-            const auto valn = itemname(ci);
-            txt(lang(valn + u8"は涸れている。"s, valn + u8" is dry."s));
-            txt(lang(
-                u8"あっ！空き瓶を井戸に落としてしまった…"s,
-                u8"Ops! You drop the empty bottle into the well..."s));
-            break;
-        }
-        cibk = ci;
-        if (inv[ci].id == 602)
-        {
-            --gdata_holy_well_count;
-            flt();
-            int stat = itemcreate(0, 516, -1, -1, 0);
-            if (stat != 0)
-            {
-                inv[ci].curse_state = curse_state_t::blessed;
-            }
-        }
-        else
-        {
-            inv[ci].param1 -= 3;
-            flt(20);
-            flttypemajor = 52000;
-            itemcreate(0, 0, -1, -1, 0);
-        }
-        txt(lang(
-            u8"空き瓶に水をすくった。"s,
-            u8"You draw water from the well into the empty bottle."s));
-        txtef(2);
-        txt(lang(
-            itemname(ci, 1) + u8"を手に入れた。"s,
-            u8"You get "s + itemname(ci, 1) + u8"."s));
-        item_stack(0, ci, 1);
-        item_stack(0, ci);
-        ci = cibk;
-        snd(17);
-        break;
-    }
-
-    item_stack(0, ci);
-    if (inv[ci].body_part != 0)
-    {
-        create_pcpic(0, true);
-    }
-    if (inv_getowner(ci) == -1)
-    {
-        cell_refresh(inv[ci].position.x, inv[ci].position.y);
-    }
-    refresh_character(0);
-    return;
-}
 
 
 
@@ -56703,7 +53468,7 @@ turn_result_t do_use_command()
     }
     if (the_item_db[inv[ci].id]->subcategory == 59500)
     {
-        return crafting_menu();
+        return blending_menu();
     }
     if (the_item_db[inv[ci].id]->subcategory == 60004)
     {
@@ -56734,7 +53499,7 @@ turn_result_t do_use_command()
         prodtype = inv[ci].function;
         snd(26);
         invctrl = 0;
-        label_18552();
+        crafting_menu();
         return turn_result_t::turn_end;
     }
     if (ibit(10, ci))
