+++ init.cpp
@@ -1,11 +1,13 @@
 #include "ability.hpp"
 #include "autopick.hpp"
+#include "blending.hpp"
 #include "buff.hpp"
 #include "cat.hpp"
 #include "character.hpp"
 #include "class.hpp"
 #include "config.hpp"
 #include "ctrl_file.hpp"
+#include "crafting.hpp"
 #include "elona.hpp"
 #include "filesystem.hpp"
 #include "fish.hpp"
