+++ input.cpp
@@ -1,5 +1,6 @@
 #include "input.hpp"
 #include "audio.hpp"
+#include "blending.hpp"
 #include "config.hpp"
 #include "elona.hpp"
 #include "variables.hpp"
