+++ main.hpp
@@ -18,15 +18,10 @@ extern int prm_853;
 
 extern int tmp_at_m172;
 
-extern elona_vector1<int> inhlist_at_m184;
 extern elona_vector1<int> inhlist;
 
 extern int i_at_m68;
 extern double r_at_m133;
-
-
-#define window_recipe(a, b, c, d, e, f) window_recipe_(b, c, d, e, f)
-
 
 
 main_menu_result_t character_making_select_sex(bool = true);
