+++ variables.hpp
@@ -1,3 +1,5 @@
+#pragma once
+
 #include "elona.hpp"
 #include "position.hpp"
 #include "enums.hpp"
@@ -25,6 +27,14 @@ ELONA_EXTERN(int anic);
 ELONA_EXTERN(elona_vector1<int> ap);
 ELONA_EXTERN(int wishfilter);
 ELONA_EXTERN(int nooracle);
+
+// crafting.cpp / blending.cpp
+ELONA_EXTERN(elona_vector1<int> rpsourcelist);
+ELONA_EXTERN(int prodtype);
+ELONA_EXTERN(elona_vector2<int> matref);
+ELONA_EXTERN(elona_vector2<int> matspot);
+ELONA_EXTERN(elona_vector1<std::string> matname);
+ELONA_EXTERN(elona_vector1<std::string> matdesc);
 
 ELONA_EXTERN(elona_vector1<int> _randcolor);
 ELONA_EXTERN(elona_vector1<int> aniref);
@@ -556,9 +566,6 @@ int addbuff(int = 0, int = 0, int = 0, int = 0);
 int advfavoriteskill(int = 0);
 int advfavoritestat(int = 0);
 int ai_check();
-int blendcheckext(int = 0, int = 0);
-int blendcheckmat(int = 0);
-int blendmatnum(int = 0, int = 0);
 int breath_list();
 int calcincome(int = 0);
 int calcmagiccontrol(int = 0, int = 0);
@@ -619,6 +626,7 @@ int get_freechara();
 int get_route(int = 0, int = 0, int = 0, int = 0);
 int getdecksum(int = 0);
 int getholdersum(int = 0);
+void getinheritance(int, elona_vector1<int>&, int&);
 int getspotsum(int = 0);
 int getunid(int = 0);
 int getworker(int = 0, int = 0);
@@ -641,7 +649,6 @@ int itemfind(int = 0, int = 0, int = 0);
 int itemusingfind(int, bool = false);
 int key_direction();
 int access_item_db(int);
-int get_required_craft_materials();
 int voting_box();
 int try_to_cast_spell();
 int try_to_reveal();
@@ -729,7 +736,6 @@ int net_read(int = 0);
 int net_send(const std::string&, int = 0);
 int opencard2(int = 0, int = 0);
 int pileremain();
-int prodcheck();
 int putcard(int = 0, int = 0);
 int randattb();
 int random_material(int = 0, int = 0);
@@ -744,7 +750,6 @@ int roll(int, int, int);
 int roll_max(int, int, int);
 int roundmargin(int = 0, int = 0);
 int route_info(int&, int&, int = 0);
-int rpdiff(int = 0, int = 0, int = 0);
 int servecard(int = 0);
 int skillexp(int = 0, int = 0, int = 0, int = 0, int = 0);
 int skillmod(int = 0, int = 0, int = 0);
@@ -813,9 +818,6 @@ std::string random_title(int = 0);
 std::string randomname();
 std::string ranktitle(int = 0);
 std::string refchara_str(int = 0, int = 0);
-std::string rpmatname(int = 0);
-std::string rpname(int = 0);
-std::string rpsuccessrate(int = 0);
 std::string sncnv(const std::string&);
 std::string trim_item_description(const std::string&, bool);
 std::string txtbuilding(int = 0, int = 0);
@@ -871,7 +873,6 @@ void chara_vanquish(int = 0);
 void chara_vomit(int = 0);
 void check_kill(int = 0, int = 0);
 void check_quest();
-void clear_rprefmat();
 void cnv_filestr(std::string&);
 void cnvbonus(int = 0, int = 0);
 void colorres(int = 0);
@@ -979,7 +980,6 @@ void change_item_material();
 void apply_item_material();
 void set_material_specific_attributes();
 void itemname_additional_info();
-void initialize_craft_material_data();
 void get_random_item_id();
 void set_item_info();
 void label_1346();
@@ -1153,7 +1153,6 @@ void label_1842();
 void label_1844();
 void label_18462();
 void label_1848();
-void label_18552();
 menu_result menu_materials();
 void casino_dealer();
 void casino_choose_card();
@@ -1176,9 +1175,6 @@ turn_result_t do_pray();
 turn_result_t do_offer();
 void label_1892();
 void label_1901();
-void initialize_recipememory();
-void initialize_recipe();
-turn_result_t crafting_menu();
 void label_1933();
 void label_19342();
 void label_1935();
@@ -1479,7 +1475,6 @@ void wake_up();
 void wet(int = 0, int = 0);
 void window(int = 0, int = 0, int = 0, int = 0, int = 0, int = 0);
 void window2(int = 0, int = 0, int = 0, int = 0, int = 0, int = 0);
-void window_recipe2(int = 0);
 void windowanime(int = 0, int = 0, int = 0, int = 0, int = 0, int = 0);
 void windowanimecorner(int = 0, int = 0, int = 0, int = 0, int = 0, int = 0);
 void cs_listbk();
